Plot accumulation curves for phyloseq-class object

Plot accumulation curves for phyloseq-class object

accu_plot(
  physeq,
  fact = NULL,
  nbSeq = TRUE,
  step = NULL,
  by.fact = FALSE,
  ci.col = NULL,
  col = NULL,
  lwd = 3,
  leg = TRUE,
  printSamplesNames = FALSE,
  CI = 2,
  ...
)

Arguments

physeq
(required): a phyloseq-class object.

fact
(required): Name of the factor in physeq@sam_data used to plot different lines

nbSeq
(logical): Either plot accumulation curves using sequences or using samples

step
(integer): distance among points calculated to plot lines. A low value give better plot but is more time consuming. Only use if nbSeq = TRUE

by.fact
(logical): First merge the OTU table by factor to plot only one line by factor

ci.col
: Color vector for confidence intervall. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to cange plot.

col
: Color vector for lines. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to change plot.

lwd
(default = 3): thickness for lines. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to cange plot.

leg
(logical): Plot legend or not. Only use if nbSeq = FALSE. If nbSeq = TRUE, you can use ggplot to cange plot.

printSamplesNames
(logical): Print samples names or not? Only use if nbSeq = TRUE.

CI
(default = 2) : Confidence intervall value used to multiply the standard error to plot confidence intervall

...
Additional arguments passed on to ggplot if nbSeq = TRUE or plot if nbSeq = FALSE

data("GlobalPatterns") GlobalPatterns_Archaea <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == 'Archaea') accu_plot(GlobalPatterns_Archaea, 'SampleType', nbSeq = TRUE, by.fact = TRUE)

Value

A ggplot2 plot representing the richness accumulation plot if nbSeq = TRUE, else, if nbSeq = FALSE return a base plot.

See also