Add information about Guild for FUNGI the FUNGuild databse

Please cite Nguyen et al. 2016 (doi:10.1016/j.funeco.2015.06.006 )

Usage

add_funguild_info(
  physeq,
  taxLevels = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  db_url = "http://www.stbates.org/funguild_db_2.php"
)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

taxLevels

Name of the 7 columns in tax_table required by funguild

db_url

a length 1 character string giving the URL to retrieve the database from

Value

A new object of class physeq with Guild information added to tax_table slot

Details

This function is mainly a wrapper of the work of others. Please make a reference to FUNGuildR package and the associate publication (doi:10.1016/j.funeco.2015.06.006 ) if you use this function.

See also

Author

Adrien Taudière

Examples

# \donttest{
if (requireNamespace("httr")) {
  d_fung_mini <- add_funguild_info(data_fungi_mini,
    taxLevels = c(
      "Domain",
      "Phylum",
      "Class",
      "Order",
      "Family",
      "Genus",
      "Species"
    )
  )
  sort(table(d_fung_mini@tax_table[, "guild"]), decreasing = TRUE)
}
#> 
#>                                                      Wood Saprotroph 
#>                                                                   11 
#>                               Undefined Saprotroph-|Wood Saprotroph| 
#>                                                                    7 
#>                                 Undefined Saprotroph-Wood Saprotroph 
#>                                                                    6 
#>                 Ectomycorrhizal-Fungal Pathogen-Undefined Saprotroph 
#>                                                                    4 
#>    Endophyte-Plant Saprotroph-Undefined Saprotroph-|Wood Saprotroph| 
#>                                                                    2 
#>                                       Plant Parasite-Wood Saprotroph 
#>                                                                    2 
#>                                      Ectomycorrhizal-Fungal Parasite 
#>                                                                    1 
#> Ectomycorrhizal-Fungal Parasite-Soil Saprotroph-Undefined Saprotroph 
#>                                                                    1 
#>                   Ectomycorrhizal-|Plant Saprotroph|-Wood Saprotroph 
#>                                                                    1 
#>                                 Fungal Parasite-Undefined Saprotroph 
#>                                                                    1 
#>                       Lichen Parasite-Plant Pathogen-Wood Saprotroph 
#>                                                                    1 
#>                                                       Plant Pathogen 
#>                                                                    1 
#>                                       Plant Pathogen-Wood Saprotroph 
#>                                                                    1 
#>                                                    |Wood Saprotroph| 
#>                                                                    1 
# }