Compute paired distances among matrix (e.g. otu_table)

Compute paired distances among matrix (e.g. otu_table)

dist_bycol(x, y, method = "bray", nperm = 99, ...)

Arguments

x
(required): A first matrix.

y
(required): A second matrix.

method
(default = 'bray'): the method to use internally in the vegdist function.

nperm
(default = 99): The number of permutations

...
(optional): others argument for `vegan::vegdist` function

Value

A list of length two : (i) a vector of observed distance ($obs) and (ii) a matrix of the distance after randomization ($null)

Note

the first column of the first matrix is compare to the first column of the second matrix, the second column of the first matrix is compare to the second column of the second matrix and so on.

See also