Filter indesirable taxa using blast against a against a custom database.

Usage

filter_asv_blast(
  physeq,
  fasta_for_db = NULL,
  database = NULL,
  clean_pq = TRUE,
  blastpath = NULL,
  id_cut = 80,
  bit_score_cut = 150,
  min_cover_cut = 50,
  add_info_to_taxtable = TRUE,
  nproc = 1
)

filter_asv_blast(
  physeq,
  fasta_for_db = NULL,
  database = NULL,
  clean_pq = TRUE,
  blastpath = NULL,
  id_cut = 80,
  bit_score_cut = 150,
  min_cover_cut = 50,
  add_info_to_taxtable = TRUE,
  nproc = 1
)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

fasta_for_db

path to a fasta file to make the blast database

database

path to a blast database

clean_pq

(logical) If set to TRUE, empty samples and empty ASV are discarded after filtering.

blastpath

path to blast program

id_cut

(default: 80) cut of in identity percent to keep ASV

bit_score_cut

(default: 150) cut of in bit score to keep result

min_cover_cut

(default: 50) cut of in query cover (%) to keep result

nproc

(default: 1) Set to number of cpus/processors to use for blast (args -num_threads for blastn command)

...

Others options for the balst_pq() function. See ?blast_pq. Note that params unique_per_seq and score_filter must be set to TRUE.

add_info_to_taxtable:

add some info from blast query to taxtable of the new physeq object. Only the information ("Query name", "Taxa name", "bit score", "% id. match", "Query cover", "e-value") for higher e-value hit. for each ASV is add to taxtable. Note that query name may be different from final taxa names as some function proposed to change the ASV names.

Value

A new phyloseq-class object.

A new phyloseq-class object.