Make Krona files using [KronaTools](https://github.com/marbl/Krona/wiki).

`r lifecycle::badge("maturing")`

Need the installation of kronatools on the computer ([installation instruction](https://github.com/marbl/Krona/wiki/Installing)). Function merge_krona allows merging multiple html files in one interactive krona file

krona(
  physeq,
  file = "krona.html",
  nb_seq = TRUE,
  ranks = "All",
  add_unassigned_rank = 0,
  name = NULL
)

Arguments

physeq

(required): a phyloseq-class object.

file

(required): the location of the html file to save

nb_seq

(logical, default to TRUE): If true, Krona set the distribution of sequences in the taxonomy. If False, Krona set the distribution of OTUs in the taxonomy.

ranks

(default="All"): Number of the taxonomic ranks to plot (num of the column in tax_table of your physeq object). Default setting plot all the ranks (argument 'All').

add_unassigned_rank

(number; default = 0). Add unassigned for rank inferior to 'add_unassigned_rank' when necessary

name

(default=NULL) : A name for intermediary files, usefull to name your krona dataset when merge using merge_krona

Value

A html file

Author

Adrien Taudière

Examples

data("GlobalPatterns")
GA <- subset_taxa(GlobalPatterns, Phylum=="Acidobacteria")
# krona(GA, "Number.of.sequences.html")
# krona(GA, "Number.of.OTUs.html", nb_seq = F)
# merge_krona(c("Number.of.sequences.html", "Number.of.OTUs.html"))