Lulu reclustering of class physeq

See https://www.nature.com/articles/s41467-017-01312-x for more information on the method.

Usage

lulu_pq(
  physeq,
  nproc = 1,
  id = 0.84,
  vsearchpath = "vsearch",
  verbose = FALSE,
  clean_pq = FALSE,
  keep_temporary_files = FALSE
)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

nproc

(default 1) Set to number of cpus/processors to use for the clustering

id

(default: 0.84) id for --usearch_global.

vsearchpath

path to vsearch.

verbose

(logical) if true, print some additional messages.

clean_pq

(logical) if true, empty samples and empty ASV are discarded before clustering.

keep_temporary_files

(logical, default: FALSE) Do we keep temporary files

Value

a list of for object

"new_physeq": The new phyloseq object (class physeq)

"discrepancy_vector": A vector of discrepancy showing for each taxonomic level the proportion of identic value before and after lulu reclustering. A value of 0.6 stands for 60% of ASV before re-clusering have identical value after re-clustering. In other word, 40% of ASV are assigned to a different taxonomic value. NA value are not counted as discrepancy.

"res_lulu": A list of the result from the lulu function

"merged_ASV": the data.frame used to merged ASV

Details

The version of LULU is a fork of Adrien Taudière (https://github.com/adrientaudiere/lulu) from https://github.com/tobiasgf/lulu

References

LULU : https://github.com/adrientaudiere/lulu forked from https://github.com/tobiasgf/lulu.

VSEARCH can be downloaded from https://github.com/torognes/vsearch.

Author

Tobias Guldberg Frøslev tobiasgf@snm.ku.dk & Adrien Taudière adrien.taudiere@zaclys.net

Examples

if (FALSE) {
data(data_fungi_sp_known)
lulu_pq(data_fungi_sp_known)
}