Plot OTU circle for phyloseq-class object

Plot OTU circle for phyloseq-class object

otu_circle(
  physeq = NULL,
  fact = NULL,
  taxa = "Order",
  nb_seq = TRUE,
  rarefy = FALSE,
  min_prop_tax = 0.01,
  min_prop_mod = 0.1,
  gap.degree = NULL,
  start.degree = NULL,
  row.col = NULL,
  grid.col = NULL,
  log10trans = F,
  ...
)

Arguments

physeq
(required): a phyloseq-class object.

fact
(required): Name of the factor to cluster samples by modalities. Need to be in physeq@sam_data.

taxa
(Default:'Order'): Name of the taxonomic rank of interest

nb_seq
(Default: TRUE): Represent the number of sequences or the number of OTUs (nb_seq = FALSE)

rarefy
(logical): Does each samples modalities need to be rarefy in order to compare them with the same amount of sequences?

min_prop_tax
(Default: 0.01): The minimum proportion for taxon to be ploted

min_prop_mod
(Default: 0.1) : The minimum proportion for modalities to be ploted

gap.degree
: Gap between two neighbour sectors. It can be a single value or a vector. If it is a vector, the first value corresponds to the gap after the first sector.

start.degree
: The starting degree from which the circle begins to draw. Note this degree is measured in the standard polar coordinate which means it is always reverse-clockwise.

row.col
: Color vector for row

grid.col
: Grid colors which correspond to sectors. The length of the vector should be either 1 or the number of sectors. It's preferred that grid.col is a named vector of which names correspond to sectors. If it is not a named vector, the order of grid.col corresponds to order of sectors.

log10trans
(logical): Should sequence be log10 transformed (more precisely by log10(1+x))?

...
Additional arguments passed on to chordDiagram or circos.par

data("GlobalPatterns") # GlobalPatterns_Archaea <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == 'Archaea') # otu_circle(GlobalPatterns_Archaea, 'SampleType')

Value

A chordDiagram plot representing the distribution of OTUs or sequences in the different modalities of the factor fact

See also