Plot DESeq2 results for a phyloseq or a DESeq2 object.

plot_deseq2_phyloseq(
  data,
  contrast = NULL,
  tax_table = NULL,
  alpha = 0.01,
  taxa = "Genus",
  select_taxa = "No",
  color_tax = "Phylum",
  tax_depth = NULL,
  verbose = TRUE,
  jitter_width = 0.1,
  ...
)

Arguments

data

(required): a phyloseq-class or a DESeqDataSet-class object.

contrast

(required):This argument specifies what comparison to extract from the object to build a results table. See results man page for more details.

tax_table

: Required if data is a DESeqDataSet-class object. The taxonomic table used to find the taxa and color_taxa arguments. If data is a phyloseq-class object, data@tax_table is used.

alpha

(default = 0.01): the significance cutoff used for optimizing the independent filtering. If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.

taxa

(default = 'Genus'): taxonomic level of interest

select_taxa

(default = 'No'): logical vector to select taxa to plot

color_tax

(default = 'Phylum'): taxonomic level used for color or a color vector.

tax_depth

(default = NULL): Taxonomic depth to test for differential distribution among contrast. If Null the analysis is done at the OTU (i.e. Species) level. If not Null data need to be a phyloseq-class object.

verbose

: whether the function print some information during the computation

jitter_width

(default=0.1) : width for the jitter positionning

...

Additional arguments passed on to DESeq or ggplot

Value

A ggplot2 plot representing DESeq2 results

See also

Author

Adrien Taudière

Examples

data("GlobalPatterns")
GP <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == "Archaea")
GP <- subset_samples(GP, GP@sam_data$SampleType %in% c("Soil", "Skin"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function '%in%': object 'GP' not found
res <- DESeq2::DESeq(phyloseq_to_deseq2(GP, ~SampleType),
  test = "Wald", fitType = "local"
)
#> converting counts to integer mode
#>   Note: levels of factors in the design contain characters other than
#>   letters, numbers, '_' and '.'. It is recommended (but not required) to use
#>   only letters, numbers, and delimiters '_' or '.', as these are safe characters
#>   for column names in R. [This is a message, not a warning or an error]
#> estimating size factors
#> Error in estimateSizeFactorsForMatrix(counts(object), locfunc = locfunc,     geoMeans = geoMeans, controlGenes = controlGenes): every gene contains at least one zero, cannot compute log geometric means
plot_deseq2_phyloseq(res, c("SampleType", "Soil", "Skin"),
  tax_table = GP@tax_table, color_tax = "Kingdom"
)
#> Error in inherits(data, "phyloseq"): object 'res' not found
plot_deseq2_phyloseq(res, c("SampleType", "Soil", "Skin"),
  tax_table = GP@tax_table, color_tax = "Kingdom",
  alpha = 0.7
)
#> Error in plot_deseq2_phyloseq(res, c("SampleType", "Soil", "Skin"), tax_table = GP@tax_table,     color_tax = "Kingdom", alpha = 0.7): object 'res' not found