Plot edgeR results for a phyloseq or a edgeR object.

Usage

plot_edgeR_pq(
  physeq,
  contrast = NULL,
  alpha = 0.01,
  taxolev = "Genus",
  color_tax = "Phylum",
  verbose = TRUE,
  ...
)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

contrast

(required):This argument specifies what comparison to extract from the object to build a results table. See results man page for more details.

alpha

(default: 0.01): the significance cutoff used for optimizing the independent filtering. If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.

taxolev

taxonomic level of interest

color_tax

taxonomic level used for color assignation

verbose

(logical): whether the function print some information during the computation

...

Additional arguments passed on to exactTest or ggplot

Value

A ggplot2 plot representing edgeR results

See also

Author

Adrien Taudière

Examples

if (FALSE) {
data(GlobalPatterns)
plot_edgeR_pq(GlobalPatterns, c("SampleType", "Soil", "Feces"),
  color_tax = "Kingdom"
)
plot_edgeR_pq(GlobalPatterns, c("SampleType", "Soil", "Feces"),
  taxolev = "Class", color_tax = "Kingdom"
)
}