Search for a list of sequence in a fasta file against physeq reference sequences using vsearch

Usage

vs_search_global(
  physeq,
  seq2search = NULL,
  path_to_fasta = NULL,
  vsearchpath = "vsearch",
  id = 0.8,
  iddef = 0,
  keep_temporary_files = FALSE
)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

seq2search

(required if path_to_fasta is NULL) Either (i) a DNAstringSet object or (ii) a character vector that will be convert to DNAstringSet using Biostrings::DNAStringSet()

path_to_fasta

(required if seq2search is NULL) a path to fasta file if seq2search is est to NULL.

vsearchpath

(default: vsearch) path to vsearch

id

(default: 0.8) id for the option --usearch_global of the vsearch software

iddef

(default: 0) iddef for the option --usearch_global of the vsearch software

keep_temporary_files

(logical, default: FALSE) Do we keep temporary files

temp.fasta (refseq in fasta)

cluster.fasta (centroid)

temp.uc (clusters)

Value

A dataframe with uc results (invisible)

Details

This function is mainly a wrapper of the work of others. Please cite vsearch.

Author

Adrien Taudière

Examples

# \donttest{
file_dna <- tempfile("dna.fa")
seqinr::write.fasta("GCCCATTAGTATTCTAGTGGGCATGCCTGTTCGAGCGTCATTTTCAACC",
  file = file_dna, names = "seq1"
)

res <- vs_search_global(data_fungi, path_to_fasta = file_dna)
unlink(file_dna)

res[res$identity != "*", ]
#>      type cluster width identity strand 6 7 cigarAlignment   query target
#> 3       H       0   312    100.0      + 0 0     64D49M199D    ASV2   seq1
#> 4       H       0   301     93.9      + 0 0     64D49M188D    ASV6   seq1
#> 8       H       0   300     85.7      + 0 0     64D49M187D   ASV10   seq1
#> 11      H       0   305     93.9      + 0 0     64D49M192D   ASV22   seq1
#> 12      H       0   315     91.8      + 0 0     64D49M202D   ASV23   seq1
#> 19      H       0   303     83.7      + 0 0     64D49M190D   ASV33   seq1
#> 20      H       0   303     93.9      + 0 0     64D49M190D   ASV31   seq1
#> 26      H       0   312     81.6      + 0 0     64D49M199D   ASV45   seq1
#> 27      H       0   295     83.7      + 0 0     64D49M182D   ASV43   seq1
#> 34      H       0   283     81.6      + 0 0     64D49M170D   ASV51   seq1
#> 35      H       0   298     81.6      + 0 0 64D43M6D6M179D   ASV52   seq1
#> 40      H       0   300     81.6      + 0 0     64D49M187D   ASV56   seq1
#> 42      H       0   294     85.7      + 0 0 64D43M6D6M175D   ASV60   seq1
#> 49      H       0   303     83.7      + 0 0 64D42M6D7M184D   ASV69   seq1
#> 50      H       0   331     83.7      + 0 0     64D49M218D   ASV65   seq1
#> 51      H       0   302     81.6      + 0 0     64D49M189D   ASV70   seq1
#> 56      H       0   304     83.7      + 0 0     64D49M191D   ASV78   seq1
#> 57      H       0   283     81.6      + 0 0     64D49M170D   ASV79   seq1
#> 59      H       0   300     81.6      + 0 0     64D49M187D   ASV80   seq1
#> 64      H       0   304     89.8      + 0 0     64D49M191D   ASV87   seq1
#> 65      H       0   300     85.7      + 0 0     64D49M187D   ASV90   seq1
#> 66      H       0   310     89.8      + 0 0     64D49M197D   ASV89   seq1
#> 68      H       0   291     83.7      + 0 0     64D49M178D   ASV92   seq1
#> 71      H       0   294     85.7      + 0 0 64D43M6D6M175D   ASV96   seq1
#> 72      H       0   330     83.7      + 0 0     65D49M216D   ASV95   seq1
#> 81      H       0   288     81.6      + 0 0     64D49M175D  ASV109   seq1
#> 82      H       0   301     89.8      + 0 0     64D49M188D  ASV106   seq1
#> 83      H       0   329     87.8      + 0 0     64D49M216D  ASV111   seq1
#> 85      H       0   343     83.7      + 0 0     65D49M229D  ASV112   seq1
#> 88      H       0   294     85.7      + 0 0 64D43M6D6M175D  ASV113   seq1
#> 90      H       0   304     85.7      + 0 0 64D43M6D6M185D  ASV119   seq1
#> 91      H       0   290     83.7      + 0 0     64D49M177D  ASV121   seq1
#> 101     H       0   311     83.7      + 0 0     64D49M198D  ASV132   seq1
#> 105     H       0   312     81.6      + 0 0     64D49M199D  ASV137   seq1
#> 112     H       0   323     85.7      + 0 0     65D49M209D  ASV146   seq1
#> 113     H       0   291     83.7      + 0 0     64D49M178D  ASV148   seq1
#> 118     H       0   303     81.6      + 0 0     64D49M190D  ASV151   seq1
#> 123     H       0   298     81.6      + 0 0 64D43M6D6M179D  ASV159   seq1
#> 125     H       0   306     83.7      + 0 0 65D42M6D7M186D  ASV163   seq1
#> 133     H       0   310     81.6      + 0 0     64D49M197D  ASV172   seq1
#> 134     H       0   321     81.6      + 0 0     64D49M208D  ASV173   seq1
#> 139     H       0   300     81.6      + 0 0     64D49M187D  ASV182   seq1
#> 140     H       0   309     89.8      + 0 0     64D49M196D  ASV183   seq1
#> 144     H       0   335     83.7      + 0 0     65D49M221D  ASV187   seq1
#> 145     H       0   292     83.7      + 0 0     64D49M179D  ASV192   seq1
#> 146     H       0   331     83.7      + 0 0     64D49M218D  ASV190   seq1
#> 147     H       0   316     93.9      + 0 0     64D49M203D  ASV194   seq1
#> 150     H       0   314     81.6      + 0 0     64D49M201D  ASV195   seq1
#> 154     H       0   295     83.7      + 0 0     64D49M182D  ASV201   seq1
#> 162     H       0   301     81.6      + 0 0     64D49M188D  ASV211   seq1
#> 165     H       0   303     89.8      + 0 0     64D49M190D  ASV215   seq1
#> 167     H       0   334     83.7      + 0 0     65D49M220D  ASV217   seq1
#> 174     H       0   302     81.6      + 0 0     64D49M189D  ASV228   seq1
#> 176     H       0   311     91.8      + 0 0     64D49M198D  ASV230   seq1
#> 179     H       0   322     85.7      + 0 0     64D49M209D  ASV233   seq1
#> 183     H       0   300     81.6      + 0 0     64D49M187D  ASV239   seq1
#> 187     H       0   318     81.6      + 0 0     64D49M205D  ASV240   seq1
#> 188     H       0   306     91.8      + 0 0  64D40MI8M194D  ASV245   seq1
#> 193     H       0   328     87.8      + 0 0     64D49M215D  ASV253   seq1
#> 195     H       0   307     93.9      + 0 0  64D9MD40M193D  ASV255   seq1
#> 197     H       0   316     81.6      + 0 0     64D49M203D  ASV256   seq1
#> 201     H       0   330     83.7      + 0 0     65D49M216D  ASV260   seq1
#> 202     H       0   291     81.6      + 0 0     64D49M178D  ASV263   seq1
#> 204     H       0   322     81.6      + 0 0     64D49M209D  ASV264   seq1
#> 209     H       0   299     83.7      + 0 0     64D49M186D  ASV271   seq1
#> 219     H       0   301     89.8      + 0 0     64D49M188D  ASV283   seq1
#> 230     H       0   301     87.8      + 0 0     64D49M188D  ASV300   seq1
#> 234     H       0   300     81.6      + 0 0     64D49M187D  ASV309   seq1
#> 236     H       0   299     87.8      + 0 0  64D40MI8M187D  ASV312   seq1
#> 238     H       0   311     89.8      + 0 0     64D49M198D  ASV313   seq1
#> 242     H       0   304     91.8      + 0 0     65D49M190D  ASV318   seq1
#> 243     H       0   331     83.7      + 0 0     64D49M218D  ASV322   seq1
#> 245     H       0   290     83.7      + 0 0     65D49M176D  ASV325   seq1
#> 247     H       0   299     83.7      + 0 0     64D49M186D  ASV327   seq1
#> 248     H       0   300     83.7      + 0 0 64D42M6D7M181D  ASV329   seq1
#> 254     H       0   288     81.6      + 0 0     64D49M175D  ASV337   seq1
#> 260     H       0   301     85.7      + 0 0     64D49M188D  ASV344   seq1
#> 261     H       0   343     83.7      + 0 0     64D49M230D  ASV342   seq1
#> 264     H       0   300     85.7      + 0 0     64D49M187D  ASV347   seq1
#> 265     H       0   305     81.6      + 0 0     64D49M192D  ASV350   seq1
#> 266     H       0   301     85.7      + 0 0     64D49M188D  ASV351   seq1
#> 267     H       0   322     81.6      + 0 0     64D49M209D  ASV349   seq1
#> 268     H       0   286     83.7      + 0 0 64D42M6D7M167D  ASV353   seq1
#> 269     H       0   304     81.6      + 0 0     64D49M191D  ASV355   seq1
#> 272     H       0   303     83.7      + 0 0 64D42M6D7M184D  ASV358   seq1
#> 274     H       0   293     81.6      + 0 0     64D49M180D  ASV362   seq1
#> 275     H       0   295     83.7      + 0 0 64D42M6D7M176D  ASV360   seq1
#> 278     H       0   331     83.7      + 0 0     64D49M218D  ASV365   seq1
#> 280     H       0   330     83.7      + 0 0     64D49M217D  ASV369   seq1
#> 282     H       0   322     81.6      + 0 0     64D49M209D  ASV371   seq1
#> 283     H       0   283     81.6      + 0 0     64D49M170D  ASV372   seq1
#> 284     H       0   303     91.8      + 0 0     64D49M190D  ASV373   seq1
#> 286     H       0   312     95.9      + 0 0     64D49M199D  ASV375   seq1
#> 289     H       0   334     83.7      + 0 0     64D49M221D  ASV378   seq1
#> 292     H       0   299     87.8      + 0 0  64D40MI8M187D  ASV381   seq1
#> 302     H       0   300     87.8      + 0 0     64D49M187D  ASV399   seq1
#> 305     H       0   298     85.7      + 0 0     64D49M185D  ASV402   seq1
#> 306     H       0   303     83.7      + 0 0 64D42M6D7M184D  ASV405   seq1
#> 307     H       0   299     81.6      + 0 0     64D49M186D  ASV406   seq1
#> 308     H       0   307     91.8      + 0 0     64D49M194D  ASV407   seq1
#> 309     H       0   322     85.7      + 0 0     64D49M209D  ASV408   seq1
#> 310     H       0   300     81.6      + 0 0     64D49M187D  ASV395   seq1
#> 311     H       0   297     85.7      + 0 0     64D49M184D  ASV409   seq1
#> 313     H       0   301     93.9      + 0 0     64D49M188D  ASV412   seq1
#> 315     H       0   294     85.7      + 0 0 64D43M6D6M175D  ASV415   seq1
#> 324     H       0   300     85.7      + 0 0     64D49M187D  ASV424   seq1
#> 327     H       0   289     81.6      + 0 0     64D49M176D  ASV427   seq1
#> 329     H       0   287     81.6      + 0 0     64D49M174D  ASV430   seq1
#> 330     H       0   295     85.7      + 0 0 64D42M6D7M176D  ASV431   seq1
#> 332     H       0   302     93.9      + 0 0     65D49M188D  ASV432   seq1
#> 338     H       0   333     83.7      + 0 0     64D49M220D  ASV438   seq1
#> 342     H       0   305     81.6      + 0 0     64D49M192D  ASV448   seq1
#> 344     H       0   305     81.6      + 0 0     64D49M192D  ASV451   seq1
#> 345     H       0   326     81.6      + 0 0     65D49M212D  ASV452   seq1
#> 346     H       0   300     81.6      + 0 0     64D49M187D  ASV453   seq1
#> 349     H       0   296     83.7      + 0 0 64D41M6D8M177D  ASV242   seq1
#> 352     H       0   294     81.6      + 0 0     64D49M181D  ASV459   seq1
#> 355     H       0   311     89.8      + 0 0     64D49M198D  ASV462   seq1
#> 358     H       0   319     83.7      + 0 0     64D49M206D  ASV463   seq1
#> 361     H       0   300     87.8      + 0 0     64D49M187D  ASV467   seq1
#> 363     H       0   304     83.7      + 0 0     64D49M191D  ASV468   seq1
#> 364     H       0   292     83.7      + 0 0     64D49M179D  ASV472   seq1
#> 367     H       0   284     81.6      + 0 0     65D49M170D  ASV473   seq1
#> 369     H       0   297     85.7      + 0 0     64D49M184D  ASV477   seq1
#> 371     H       0   304     81.6      + 0 0     64D49M191D  ASV480   seq1
#> 377     H       0   300     81.6      + 0 0     64D49M187D  ASV486   seq1
#> 383     H       0   301     85.7      + 0 0     64D49M188D  ASV492   seq1
#> 384     H       0   319     81.6      + 0 0     64D49M206D  ASV494   seq1
#> 387     H       0   289     83.7      + 0 0 64D42M6D7M170D  ASV496   seq1
#> 388     H       0   300     89.8      + 0 0     64D49M187D  ASV497   seq1
#> 389     H       0   302     81.6      + 0 0     64D49M189D  ASV499   seq1
#> 392     H       0   300     81.6      + 0 0     64D49M187D  ASV502   seq1
#> 395     H       0   296     83.7      + 0 0     64D49M183D  ASV507   seq1
#> 396     H       0   302     81.6      + 0 0     64D49M189D  ASV508   seq1
#> 400     H       0   293     81.6      + 0 0     64D49M180D  ASV511   seq1
#> 403     H       0   297     83.7      + 0 0     64D49M184D  ASV517   seq1
#> 404     H       0   291     91.8      + 0 0     64D49M178D  ASV519   seq1
#> 405     H       0   289     81.6      + 0 0 64D42M6D7M170D  ASV520   seq1
#> 406     H       0   291     85.7      + 0 0     64D49M178D  ASV521   seq1
#> 408     H       0   296     87.8      + 0 0  64D40MI8M184D  ASV524   seq1
#> 409     H       0   335     83.7      + 0 0     65D49M221D  ASV522   seq1
#> 410     H       0   329     83.7      + 0 0     64D49M216D  ASV526   seq1
#> 411     H       0   305     81.6      + 0 0     63D49M193D  ASV527   seq1
#> 417     H       0   329     81.6      + 0 0     64D49M216D  ASV533   seq1
#> 418     H       0   290     81.6      + 0 0 64D42M6D7M171D  ASV535   seq1
#> 419     H       0   310     89.8      + 0 0     64D49M197D  ASV534   seq1
#> 421     H       0   335     83.7      + 0 0     65D49M221D  ASV537   seq1
#> 422     H       0   295     83.7      + 0 0     64D49M182D  ASV542   seq1
#> 423     H       0   320     83.7      + 0 0     64D49M207D  ASV540   seq1
#> 426     H       0   300     81.6      + 0 0     64D49M187D  ASV544   seq1
#> 428     H       0   302     83.7      + 0 0     64D49M189D  ASV546   seq1
#> 434     H       0   312     91.8      + 0 0     64D49M199D  ASV541   seq1
#> 435     H       0   302     81.6      + 0 0     64D49M189D  ASV551   seq1
#> 439     H       0   303     83.7      + 0 0 64D42M6D7M184D  ASV556   seq1
#> 442     H       0   293     83.7      + 0 0     64D49M180D  ASV560   seq1
#> 443     H       0   301     81.6      + 0 0     64D49M188D  ASV561   seq1
#> 446     H       0   295     85.7      + 0 0 64D42M6D7M176D  ASV563   seq1
#> 448     H       0   292     83.7      + 0 0     64D49M179D  ASV564   seq1
#> 452     H       0   307     89.8      + 0 0     64D49M194D  ASV570   seq1
#> 453     H       0   301     81.6      + 0 0     65D49M187D  ASV569   seq1
#> 454     H       0   283     81.6      + 0 0     64D49M170D  ASV573   seq1
#> 455     H       0   301     81.6      + 0 0     65D49M187D  ASV572   seq1
#> 456     H       0   300     85.7      + 0 0     65D49M186D  ASV571   seq1
#> 457     H       0   283     81.6      + 0 0     64D49M170D  ASV574   seq1
#> 459     H       0   313    100.0      + 0 0     65D49M199D  ASV576   seq1
#> 461     H       0   304     85.7      + 0 0 64D43M6D6M185D  ASV578   seq1
#> 463     H       0   300     81.6      + 0 0     64D49M187D  ASV581   seq1
#> 465     H       0   319     81.6      + 0 0     64D49M206D  ASV582   seq1
#> 466     H       0   293     81.6      + 0 0     64D49M180D  ASV585   seq1
#> 467     H       0   291     83.7      + 0 0 64D42M6D7M172D  ASV586   seq1
#> 470     H       0   300     83.7      + 0 0     65D49M186D  ASV590   seq1
#> 474     H       0   318     83.7      + 0 0     63D49M206D  ASV593   seq1
#> 479     H       0   331     83.7      + 0 0     65D49M217D  ASV597   seq1
#> 484     H       0   301     85.7      + 0 0 65D43M6D6M181D  ASV603   seq1
#> 485     H       0   302     83.7      + 0 0     65D49M188D  ASV604   seq1
#> 486     H       0   290     83.7      + 0 0 64D42M6D7M171D  ASV608   seq1
#> 487     H       0   322     85.7      + 0 0     64D49M209D  ASV607   seq1
#> 489     H       0   292     91.8      + 0 0     64D49M179D  ASV614   seq1
#> 508     H       0   284     85.7      + 0 0     64D49M171D  ASV635   seq1
#> 509     H       0   312     87.8      + 0 0     64D49M199D  ASV634   seq1
#> 513     H       0   305     85.7      + 0 0 64D43M6D6M186D  ASV637   seq1
#> 518     H       0   291     81.6      + 0 0     65D49M177D  ASV645   seq1
#> 520     H       0   309     81.6      + 0 0     64D49M196D  ASV649   seq1
#> 522     H       0   304     83.7      + 0 0     64D49M191D  ASV652   seq1
#> 524     H       0   339     83.7      + 0 0     65D49M225D  ASV650   seq1
#> 528     H       0   322     83.7      + 0 0     64D49M209D  ASV658   seq1
#> 529     H       0   289     81.6      + 0 0     64D49M176D  ASV659   seq1
#> 530     H       0   302     85.7      + 0 0     65D49M188D  ASV661   seq1
#> 539     H       0   294     81.6      + 0 0     64D49M181D  ASV671   seq1
#> 540     H       0   301     85.7      + 0 0     65D49M187D  ASV670   seq1
#> 543     H       0   303     83.7      + 0 0 64D42M6D7M184D  ASV674   seq1
#> 545     H       0   306     81.6      + 0 0     64D49M193D  ASV676   seq1
#> 546     H       0   324     85.7      + 0 0     65D49M210D  ASV677   seq1
#> 551     H       0   346     83.7      + 0 0     65D49M232D  ASV679   seq1
#> 552     H       0   309     85.7      + 0 0     64D49M196D  ASV683   seq1
#> 554     H       0   288     81.6      + 0 0     64D49M175D  ASV684   seq1
#> 556     H       0   330     83.7      + 0 0     64D49M217D  ASV687   seq1
#> 558     H       0   333     83.7      + 0 0     65D49M219D  ASV688   seq1
#> 565     H       0   301     83.7      + 0 0     65D49M187D  ASV695   seq1
#> 568     H       0   298     83.7      + 0 0     64D49M185D  ASV700   seq1
#> 570     H       0   301     85.7      + 0 0 64D42M6D7M182D  ASV553   seq1
#> 572     H       0   300     85.7      + 0 0     64D49M187D  ASV704   seq1
#> 574     H       0   299     85.7      + 0 0     64D49M186D  ASV705   seq1
#> 575     H       0   295     83.7      + 0 0     64D49M182D  ASV710   seq1
#> 577     H       0   312     81.6      + 0 0     64D49M199D  ASV711   seq1
#> 582     H       0   311     81.6      + 0 0     64D49M198D  ASV715   seq1
#> 583     H       0   301     81.6      + 0 0     64D49M188D  ASV717   seq1
#> 584     H       0   290     85.7      + 0 0 65D42M7D7M169D  ASV720   seq1
#> 587     H       0   300     85.7      + 0 0     64D49M187D  ASV722   seq1
#> 590     H       0   313     81.6      + 0 0     64D49M200D  ASV726   seq1
#> 591     H       0   302     83.7      + 0 0     64D49M189D  ASV727   seq1
#> 593     H       0   300     93.9      + 0 0     64D49M187D  ASV729   seq1
#> 600     H       0   313     83.7      + 0 0     64D49M200D  ASV737   seq1
#> 603     H       0   305     81.6      + 0 0     64D49M192D  ASV740   seq1
#> 607     H       0   302     81.6      + 0 0     64D49M189D  ASV744   seq1
#> 611     H       0   334     81.6      + 0 0     65D49M220D  ASV746   seq1
#> 613     H       0   289     81.6      + 0 0     64D49M176D  ASV751   seq1
#> 618     H       0   301     81.6      + 0 0     64D49M188D  ASV755   seq1
#> 623     H       0   312     81.6      + 0 0     64D49M199D  ASV760   seq1
#> 625     H       0   311     81.6      + 0 0     64D49M198D  ASV765   seq1
#> 629     H       0   292     85.7      + 0 0     64D49M179D  ASV769   seq1
#> 631     H       0   308     85.7      + 0 0     64D49M195D  ASV770   seq1
#> 634     H       0   330     83.7      + 0 0     64D49M217D  ASV774   seq1
#> 636     H       0   299     87.8      + 0 0  65D40MI8M186D  ASV775   seq1
#> 637     H       0   298     89.8      + 0 0  64D40MI8M186D  ASV777   seq1
#> 639     H       0   303     83.7      + 0 0 64D42M6D7M184D  ASV776   seq1
#> 642     H       0   323     83.7      + 0 0     64D49M210D  ASV781   seq1
#> 643     H       0   304     81.6      + 0 0     65D49M190D  ASV783   seq1
#> 645     H       0   352     83.7      + 0 0     65D49M238D  ASV784   seq1
#> 647     H       0   302     83.7      + 0 0     64D49M189D  ASV786   seq1
#> 651     H       0   301     83.7      + 0 0     65D49M187D  ASV792   seq1
#> 652     H       0   294     85.7      + 0 0 64D43M6D6M175D  ASV796   seq1
#> 653     H       0   303     83.7      + 0 0 64D42M6D7M184D  ASV797   seq1
#> 656     H       0   301     83.7      + 0 0     64D49M188D  ASV799   seq1
#> 658     H       0   311     93.9      + 0 0     64D49M198D  ASV802   seq1
#> 660     H       0   296     81.6      + 0 0     64D49M183D  ASV805   seq1
#> 661     H       0   329     85.7      + 0 0     64D49M216D  ASV806   seq1
#> 671     H       0   303     81.6      + 0 0     64D49M190D  ASV820   seq1
#> 672     H       0   302     93.9      + 0 0     65D49M188D  ASV821   seq1
#> 673     H       0   301     83.7      + 0 0     65D49M187D  ASV822   seq1
#> 676     H       0   302     81.6      + 0 0     64D49M189D  ASV826   seq1
#> 677     H       0   295     83.7      + 0 0     64D49M182D  ASV827   seq1
#> 679     H       0   291     85.7      + 0 0     64D49M178D  ASV828   seq1
#> 685     H       0   305     81.6      + 0 0     65D49M191D  ASV834   seq1
#> 686     H       0   292     81.6      + 0 0     64D49M179D  ASV836   seq1
#> 688     H       0   293     83.7      + 0 0     65D49M179D  ASV837   seq1
#> 693     H       0   301     81.6      + 0 0     65D49M187D  ASV842   seq1
#> 695     H       0   305     85.7      + 0 0     64D49M192D  ASV844   seq1
#> 700     H       0   300     81.6      + 0 0     64D49M187D  ASV850   seq1
#> 701     H       0   305     81.6      + 0 0     64D49M192D  ASV852   seq1
#> 705     H       0   307     93.9      + 0 0  64D9MD40M193D  ASV854   seq1
#> 710     H       0   295     83.7      + 0 0 64D42M6D7M176D  ASV861   seq1
#> 724     H       0   302     83.7      + 0 0     65D49M188D  ASV877   seq1
#> 725     H       0   294     81.6      + 0 0     64D49M181D  ASV879   seq1
#> 726     H       0   299     81.6      + 0 0 65D43M6D6M179D  ASV880   seq1
#> 727     H       0   299     83.7      + 0 0     64D49M186D  ASV883   seq1
#> 734     H       0   301     83.7      + 0 0     64D49M188D  ASV891   seq1
#> 738     H       0   296     81.6      + 0 0     64D49M183D  ASV900   seq1
#> 739     H       0   319     83.7      + 0 0     64D49M206D  ASV901   seq1
#> 741     H       0   306     85.7      + 0 0 66D43M6D6M185D  ASV903   seq1
#> 742     H       0   291     81.6      + 0 0     64D49M178D  ASV904   seq1
#> 745     H       0   334     83.7      + 0 0     65D49M220D  ASV905   seq1
#> 746     H       0   275     83.7      + 0 0 64D42M5D7M157D  ASV908   seq1
#> 750     H       0   332     83.7      + 0 0     66D49M217D  ASV911   seq1
#> 751     H       0   304     81.6      + 0 0     64D49M191D  ASV913   seq1
#> 752     H       0   295     83.7      + 0 0     64D49M182D  ASV915   seq1
#> 753     H       0   304     83.7      + 0 0 64D42M6D7M185D  ASV916   seq1
#> 754     H       0   302     85.7      + 0 0     64D49M189D  ASV917   seq1
#> 755     H       0   295     85.7      + 0 0     64D49M182D  ASV918   seq1
#> 760     H       0   312     83.7      + 0 0     64D49M199D  ASV921   seq1
#> 762     H       0   288     93.9      + 0 0     65D49M174D  ASV925   seq1
#> 763     H       0   300     85.7      + 0 0     64D49M187D  ASV926   seq1
#> 765     H       0   302     81.6      + 0 0     64D49M189D  ASV927   seq1
#> 767     H       0   302     85.7      + 0 0 64D42M6D7M183D  ASV930   seq1
#> 770     H       0   288     81.6      + 0 0     64D49M175D  ASV934   seq1
#> 771     H       0   324     85.7      + 0 0     65D49M210D  ASV935   seq1
#> 772     H       0   300     85.7      + 0 0     64D49M187D  ASV936   seq1
#> 774     H       0   302     81.6      + 0 0     65D49M188D  ASV938   seq1
#> 775     H       0   304     81.6      + 0 0     65D49M190D  ASV939   seq1
#> 779     H       0   304     81.6      + 0 0     64D49M191D  ASV941   seq1
#> 783     H       0   286     81.6      + 0 0     66D49M171D  ASV947   seq1
#> 785     H       0   308     91.8      + 0 0  65D40MI8M195D  ASV949   seq1
#> 788     H       0   301     91.8      + 0 0     65D49M187D  ASV951   seq1
#> 790     H       0   287     85.7      + 0 0 64D42M6D7M168D  ASV955   seq1
#> 791     H       0   298     83.7      + 0 0     64D49M185D  ASV956   seq1
#> 792     H       0   302     85.7      + 0 0     64D49M189D  ASV957   seq1
#> 794     H       0   301     85.7      + 0 0     64D49M188D  ASV958   seq1
#> 797     H       0   313     87.8      + 0 0     65D49M199D  ASV962   seq1
#> 799     H       0   310     81.6      + 0 0     64D49M197D  ASV963   seq1
#> 800     H       0   301     85.7      + 0 0 64D42M6D7M182D  ASV966   seq1
#> 803     H       0   302     93.9      + 0 0     65D49M188D  ASV969   seq1
#> 805     H       0   302     85.7      + 0 0     64D49M189D  ASV970   seq1
#> 807     H       0   295     81.6      + 0 0     64D49M182D  ASV973   seq1
#> 815     H       0   312     81.6      + 0 0     64D49M199D  ASV981   seq1
#> 816     H       0   302     81.6      + 0 0     65D49M188D  ASV983   seq1
#> 818     H       0   295     85.7      + 0 0 64D43M6D6M176D  ASV984   seq1
#> 819     H       0   301     81.6      + 0 0     64D49M188D  ASV985   seq1
#> 821     H       0   323     91.8      + 0 0     65D49M209D  ASV988   seq1
#> 824     H       0   289     83.7      + 0 0     64D49M176D  ASV992   seq1
#> 827     H       0   315     85.7      + 0 0     65D49M201D  ASV995   seq1
#> 828     H       0   313     81.6      + 0 0     64D49M200D  ASV996   seq1
#> 830     H       0   332     87.8      + 0 0     64D49M219D  ASV997   seq1
#> 831     H       0   295     83.7      + 0 0     64D49M182D  ASV999   seq1
#> 833     H       0   292     81.6      + 0 0     64D49M179D ASV1001   seq1
#> 834     H       0   300     85.7      + 0 0     64D49M187D ASV1003   seq1
#> 835     H       0   332     83.7      + 0 0     65D49M218D ASV1002   seq1
#> 836     H       0   320     83.7      + 0 0     65D49M206D ASV1004   seq1
#> 838     H       0   301     87.8      + 0 0 64D42M5D7M183D ASV1006   seq1
#> 840     H       0   299     89.8      + 0 0 64D43M6D6M180D ASV1008   seq1
#> 843     H       0   294     98.0      + 0 0 64D43M6D6M175D ASV1011   seq1
#> 845     H       0   301     87.8      + 0 0     64D49M188D ASV1014   seq1
#> 846     H       0   305     85.7      + 0 0 64D43M6D6M186D ASV1015   seq1
#> 849     H       0   294     85.7      + 0 0 64D43M6D6M175D ASV1020   seq1
#> 850     H       0   294     85.7      + 0 0 64D43M6D6M175D ASV1021   seq1
#> 851     H       0   301     81.6      + 0 0     65D49M187D ASV1022   seq1
#> 852     H       0   315     81.6      + 0 0     65D49M201D ASV1023   seq1
#> 855     H       0   298     83.7      + 0 0     64D49M185D ASV1025   seq1
#> 858     H       0   312     87.8      + 0 0     64D49M199D ASV1027   seq1
#> 860     H       0   322     85.7      + 0 0     65D49M208D ASV1030   seq1
#> 864     H       0   294     85.7      + 0 0 64D43M6D6M175D ASV1036   seq1
#> 865     H       0   302     85.7      + 0 0     64D49M189D ASV1038   seq1
#> 867     H       0   296     81.6      + 0 0     64D49M183D ASV1040   seq1
#> 868     H       0   300     89.8      + 0 0     64D49M187D ASV1039   seq1
#> 874     H       0   294     83.7      + 0 0 64D42M6D7M175D ASV1046   seq1
#> 876     H       0   301     81.6      + 0 0     65D49M187D ASV1049   seq1
#> 879     H       0   292     91.8      + 0 0     64D49M179D ASV1055   seq1
#> 880     H       0   299     81.6      + 0 0     64D49M186D ASV1056   seq1
#> 882     H       0   300     89.8      + 0 0     64D49M187D ASV1058   seq1
#> 883     H       0   300     93.9      + 0 0     64D49M187D ASV1059   seq1
#> 885     H       0   313     85.7      + 0 0     65D49M199D ASV1062   seq1
#> 889     H       0   294     98.0      + 0 0 64D43M6D6M175D ASV1066   seq1
#> 891     H       0   314     81.6      + 0 0     65D49M200D ASV1067   seq1
#> 894     H       0   296     81.6      + 0 0 64D43M6D6M177D ASV1070   seq1
#> 901     H       0   289     81.6      + 0 0     65D49M175D ASV1081   seq1
#> 902     H       0   312     87.8      + 0 0     64D49M199D ASV1082   seq1
#> 904     H       0   292     93.9      + 0 0     65D49M178D ASV1084   seq1
#> 907     H       0   292     85.7      + 0 0     65D49M178D ASV1090   seq1
#> 908     H       0   311     81.6      + 0 0     64D49M198D ASV1087   seq1
#> 909     H       0   302     83.7      + 0 0     64D49M189D ASV1091   seq1
#> 912     H       0   301     81.6      + 0 0     64D49M188D ASV1095   seq1
#> 918     H       0   313     89.8      + 0 0     66D49M198D ASV1101   seq1
#> 919     H       0   302     93.9      + 0 0     65D49M188D ASV1102   seq1
#> 920     H       0   303     85.7      + 0 0 65D42M6D7M183D ASV1103   seq1
#> 924     H       0   304     83.7      + 0 0     64D49M191D ASV1109   seq1
#> 925     H       0   303     81.6      + 0 0     66D49M188D ASV1110   seq1
#> 927     H       0   318     81.6      + 0 0     65D49M204D ASV1108   seq1
#> 929     H       0   296     87.8      + 0 0 65D43M6D6M176D ASV1114   seq1
#> 931     H       0   322     85.7      + 0 0     65D49M208D ASV1115   seq1
#> 932     H       0   302     83.7      + 0 0     64D49M189D ASV1117   seq1
#> 934     H       0   299     81.6      + 0 0     64D49M186D ASV1120   seq1
#> 936     H       0   307     91.8      + 0 0  65D40MI8M194D ASV1121   seq1
#> 937     H       0   295     83.7      + 0 0     64D49M182D ASV1124   seq1
#> 938     H       0   300     81.6      + 0 0     64D49M187D ASV1125   seq1
#> 940     H       0   295     85.7      + 0 0 64D43M6D6M176D ASV1127   seq1
#> 946     H       0   330     83.7      + 0 0     65D49M216D ASV1133   seq1
#> 947     H       0   305     85.7      + 0 0 64D43M6D6M186D ASV1134   seq1
#> 953     H       0   315     83.7      + 0 0     64D49M202D ASV1141   seq1
#> 957     H       0   299     85.7      + 0 0 64D43M6D6M180D ASV1148   seq1
#> 958     H       0   302     93.9      + 0 0     64D49M189D ASV1150   seq1
#> 959     H       0   308     89.8      + 0 0     64D49M195D ASV1152   seq1
#> 960     H       0   287     85.7      + 0 0     64D49M174D ASV1154   seq1
#> 962     H       0   295     85.7      + 0 0 64D41M6D8M176D ASV1156   seq1
#> 963     H       0   290     81.6      + 0 0     65D49M176D ASV1157   seq1
#> 964     H       0   299     85.7      + 0 0     64D49M186D ASV1158   seq1
#> 967     H       0   300     81.6      + 0 0     64D49M187D ASV1160   seq1
#> 968     H       0   307     81.6      + 0 0     65D49M193D ASV1161   seq1
#> 969     H       0   312     85.7      + 0 0     65D49M198D ASV1163   seq1
#> 972     H       0   305     83.7      + 0 0     65D49M191D ASV1165   seq1
#> 974     H       0   325     85.7      + 0 0     66D49M210D ASV1168   seq1
#> 976     H       0   300     85.7      + 0 0     64D49M187D ASV1172   seq1
#> 977     H       0   299     81.6      + 0 0 64D42M6D7M180D ASV1173   seq1
#> 978     H       0   315     81.6      + 0 0     65D49M201D ASV1167   seq1
#> 979     H       0   335     83.7      + 0 0     64D49M222D ASV1174   seq1
#> 981     H       0   303     83.7      + 0 0     65D49M189D ASV1175   seq1
#> 983     H       0   334     83.7      + 0 0     66D49M219D ASV1177   seq1
#> 986     H       0   311     83.7      + 0 0 64D42M6D7M192D ASV1181   seq1
#> 989     H       0   296     85.7      + 0 0 65D43M6D6M176D ASV1186   seq1
#> 990     H       0   322     83.7      + 0 0     64D49M209D ASV1185   seq1
#> 991     H       0   283     81.6      + 0 0     64D49M170D ASV1187   seq1
#> 992     H       0   293     85.7      + 0 0     64D49M180D ASV1188   seq1
#> 994     H       0   296     83.7      + 0 0 65D42M6D7M176D ASV1189   seq1
#> 996     H       0   304     81.6      + 0 0     65D49M190D ASV1193   seq1
#> 1001    H       0   295     83.7      + 0 0     64D49M182D ASV1199   seq1
#> 1003    H       0   326     83.7      + 0 0     65D49M212D ASV1200   seq1
#> 1005    H       0   312     83.7      + 0 0     65D49M198D ASV1205   seq1
#> 1006    H       0   316     83.7      + 0 0     64D49M203D ASV1206   seq1
#> 1010    H       0   291     91.8      + 0 0     65D49M177D ASV1029   seq1
#> 1012    H       0   336     81.6      + 0 0     65D49M222D ASV1211   seq1
#> 1017    H       0   302     83.7      + 0 0     65D49M188D ASV1219   seq1
#> 1019    H       0   357     85.7      + 0 0     65D49M243D ASV1221   seq1
#> 1023    H       0   286     85.7      + 0 0 64D42M6D7M167D ASV1226   seq1
#> 1024    H       0   334     83.7      + 0 0     65D49M220D ASV1227   seq1
#> 1026    H       0   294     81.6      + 0 0 64D42M6D7M175D ASV1229   seq1
#> 1027    H       0   334     83.7      + 0 0     64D49M221D ASV1230   seq1
#> 1028    H       0   304     89.8      + 0 0     64D49M191D ASV1228   seq1
#> 1031    H       0   301     81.6      + 0 0     64D49M188D ASV1233   seq1
#> 1032    H       0   319     83.7      + 0 0     64D49M206D ASV1213   seq1
#> 1039    H       0   294     83.7      + 0 0 64D43M6D6M175D ASV1243   seq1
#> 1043    H       0   301     85.7      + 0 0     64D49M188D ASV1252   seq1
#> 1044    H       0   315     83.7      + 0 0     66D49M200D ASV1247   seq1
#> 1045    H       0   306     83.7      + 0 0     65D49M192D ASV1253   seq1
#> 1047    H       0   305     91.8      + 0 0     65D49M191D ASV1255   seq1
#> 1049    H       0   314     83.7      + 0 0     66D49M199D ASV1258   seq1
#> 1056    H       0   330     87.8      + 0 0     64D49M217D ASV1265   seq1
#> 1065    H       0   300     81.6      + 0 0 64D41M6D8M181D ASV1274   seq1
#> 1067    H       0   299     98.0      + 0 0 65D43M6D6M179D ASV1276   seq1
#> 1069    H       0   284     81.6      + 0 0     65D49M170D ASV1279   seq1
#> 1071    H       0   299     83.7      + 0 0     64D49M186D ASV1281   seq1
#> 1072    H       0   296     85.7      + 0 0 64D42M6D7M177D ASV1282   seq1
#> 1074    H       0   312    100.0      + 0 0     64D49M199D ASV1284   seq1
#> 1075    H       0   305     89.8      + 0 0     64D49M192D ASV1285   seq1
#> 1076    H       0   294     87.8      + 0 0     65D49M180D ASV1286   seq1
#> 1077    H       0   302     85.7      + 0 0 64D42M6D7M183D ASV1287   seq1
#> 1078    H       0   297     83.7      + 0 0     65D49M183D ASV1288   seq1
#> 1079    H       0   305     81.6      + 0 0     65D49M191D ASV1289   seq1
#> 1080    H       0   294     85.7      + 0 0 64D43M6D6M175D ASV1290   seq1
#> 1081    H       0   290     81.6      + 0 0     64D49M177D ASV1291   seq1
#> 1083    H       0   330     83.7      + 0 0     64D49M217D ASV1293   seq1
#> 1085    H       0   301     89.8      + 0 0     64D49M188D ASV1296   seq1
#> 1087    H       0   299     83.7      + 0 0     64D49M186D ASV1297   seq1
#> 1091    H       0   332     83.7      + 0 0     65D49M218D ASV1303   seq1
#> 1094    H       0   294     81.6      + 0 0     64D49M181D ASV1310   seq1
#> 1095    H       0   300     81.6      + 0 0     64D49M187D ASV1311   seq1
#> 1096    H       0   298     81.6      + 0 0     65D49M184D ASV1301   seq1
#> 1097    H       0   325     85.7      + 0 0     66D49M210D ASV1312   seq1
#> 1098    H       0   305     83.7      + 0 0     64D49M192D ASV1313   seq1
#> 1101    H       0   352     83.7      + 0 0     65D49M238D ASV1315   seq1
#> 1103    H       0   293    100.0      + 0 0     65D49M179D ASV1319   seq1
#> 1109    H       0   312     89.8      + 0 0     64D49M199D ASV1323   seq1
#> 1114    H       0   291     83.7      + 0 0     65D49M177D ASV1331   seq1
#> 1116    H       0   330     83.7      + 0 0     65D49M216D ASV1332   seq1
#> 1128    H       0   303     81.6      + 0 0     65D49M189D ASV1345   seq1
#> 1129    H       0   287     81.6      + 0 0     64D49M174D ASV1351   seq1
#> 1131    H       0   302     81.6      + 0 0     65D49M188D ASV1352   seq1
#> 1134    H       0   283     81.6      + 0 0     64D49M170D ASV1356   seq1
#> 1138    H       0   284     81.6      + 0 0     64D49M171D ASV1361   seq1
#> 1139    H       0   301     85.7      + 0 0     65D49M187D ASV1362   seq1
#> 1143    H       0   294     85.7      + 0 0 64D43M6D6M175D ASV1366   seq1
#> 1145    H       0   291     83.7      + 0 0 64D42M6D7M172D ASV1368   seq1
#> 1146    H       0   312     83.7      + 0 0     64D49M199D ASV1369   seq1
#> 1147    H       0   290     85.7      + 0 0     64D49M177D ASV1370   seq1
#> 1148    H       0   312     81.6      + 0 0     64D49M199D ASV1371   seq1
#> 1149    H       0   326     85.7      + 0 0     65D49M212D ASV1372   seq1
#> 1155    H       0   295     81.6      + 0 0     64D49M182D ASV1379   seq1
#> 1156    H       0   322     85.7      + 0 0     64D49M209D ASV1378   seq1
#> 1157    H       0   296     85.7      + 0 0 65D43M6D6M176D ASV1380   seq1
#> 1158    H       0   295     81.6      + 0 0     64D49M182D ASV1384   seq1
#> 1159    H       0   312     81.6      + 0 0     64D49M199D ASV1381   seq1
#> 1163    H       0   301     81.6      + 0 0     64D49M188D ASV1386   seq1
#> 1164    H       0   297     83.7      + 0 0     65D49M183D ASV1387   seq1
#> 1166    H       0   308     93.9      + 0 0  64D9MD40M194D ASV1389   seq1
#> 1168    H       0   294     81.6      + 0 0     64D49M181D ASV1393   seq1
#> 1169    H       0   292     81.6      + 0 0     65D49M178D ASV1394   seq1
#> 1171    H       0   305     83.7      + 0 0     65D49M191D ASV1396   seq1
#> 1176    H       0   299     83.7      + 0 0     64D49M186D ASV1401   seq1
#> 1177    H       0   318     81.6      + 0 0     65D49M204D ASV1402   seq1
#> 1178    H       0   313     81.6      + 0 0     64D49M200D ASV1403   seq1
#> 1179    H       0   323     81.6      + 0 0     65D49M209D ASV1404   seq1
#> 1180    H       0   289     81.6      + 0 0     64D49M176D ASV1405   seq1
#> 1187    H       0   298     83.7      + 0 0     64D49M185D ASV1414   seq1
#> 1189    H       0   312     81.6      + 0 0     64D49M199D ASV1415   seq1
#> 1192    H       0   299     83.7      + 0 0     65D49M185D ASV1419   seq1
#> 1194    H       0   305     81.6      + 0 0     63D49M193D ASV1422   seq1
#> 1195    H       0   347     83.7      + 0 0     65D49M233D ASV1421   seq1
#> 1197    H       0   330     87.8      + 0 0     64D49M217D ASV1416   seq1
#> 1205    H       0   295     85.7      + 0 0 64D42M6D7M176D ASV1433   seq1
#> 1206    H       0   332     87.8      + 0 0     64D49M219D ASV1432   seq1
#> 1207    H       0   319     81.6      + 0 0     65D49M205D ASV1435   seq1
#> 1213    H       0   300     81.6      + 0 0     65D49M186D ASV1444   seq1
#> 1216    H       0   301     83.7      + 0 0     65D49M187D ASV1448   seq1
#> 1219    H       0   333     83.7      + 0 0     65D49M219D ASV1452   seq1
#> 1221    H       0   295     85.7      + 0 0 64D43M6D6M176D ASV1455   seq1
#> 1223    H       0   322     83.7      + 0 0     65D49M208D ASV1454   seq1
#> 1224    H       0   290     83.7      + 0 0     64D49M177D ASV1457   seq1
#> 1226    H       0   324     83.7      + 0 0     66D49M209D ASV1458   seq1
#> 1227    H       0   303     93.9      + 0 0     64D49M190D ASV1459   seq1
#> 1233    H       0   340     85.7      + 0 0 64D43M5D6M222D ASV1467   seq1
#> 1234    H       0   295     81.6      + 0 0     64D49M182D ASV1468   seq1
#> 1238    H       0   298     83.7      + 0 0 64D42M6D7M179D ASV1472   seq1
#> 1239    H       0   303     81.6      + 0 0     65D49M189D ASV1476   seq1
#> 1242    H       0   320     83.7      + 0 0     63D49M208D ASV1478   seq1
#> 1244    H       0   300     81.6      + 0 0     65D49M186D ASV1481   seq1
#> 1246    H       0   304     83.7      + 0 0     64D49M191D ASV1482   seq1
#> 1247    H       0   311     81.6      + 0 0     64D49M198D ASV1484   seq1
#> 1249    H       0   291     81.6      + 0 0     64D49M178D ASV1486   seq1
#> 1256    H       0   301     81.6      + 0 0     64D49M188D ASV1492   seq1
#> 1257    H       0   295     81.6      + 0 0     64D49M182D ASV1495   seq1
#> 1259    H       0   301     83.7      + 0 0     64D49M188D ASV1497   seq1
#> 1262    H       0   303     81.6      + 0 0     65D49M189D ASV1501   seq1
#> 1265    H       0   293     81.6      + 0 0     64D49M180D ASV1504   seq1
#> 1266    H       0   313     85.7      + 0 0     65D49M199D ASV1507   seq1
#> 1271    H       0   295     81.6      + 0 0     64D49M182D ASV1515   seq1
#> 1275    H       0   341     83.7      + 0 0     66D49M226D ASV1521   seq1
#> 1277    H       0   300     81.6      + 0 0     64D49M187D ASV1524   seq1
#> 1278    H       0   286     85.7      + 0 0 64D43M6D6M167D ASV1526   seq1
#> 1284    H       0   295     85.7      + 0 0 64D43M6D6M176D ASV1532   seq1
#> 1286    H       0   300     85.7      + 0 0     64D49M187D ASV1534   seq1
#> 1297    H       0   310     85.7      + 0 0     64D49M197D ASV1548   seq1
#> 1299    H       0   302     85.7      + 0 0     65D49M188D ASV1549   seq1
#> 1301    H       0   323     83.7      + 0 0     64D49M210D ASV1552   seq1
#> 1304    H       0   323     85.7      + 0 0     65D49M209D ASV1558   seq1
#> 1305    H       0   318     89.8      + 0 0     65D49M204D ASV1557   seq1
#> 1307    H       0   308     85.7      + 0 0     65D49M194D ASV1559   seq1
#> 1310    H       0   326     83.7      + 0 0     65D49M212D ASV1562   seq1
#> 1333    H       0   311     85.7      + 0 0 64D42M6D7M192D ASV1589   seq1
#> 1337    H       0   317     85.7      + 0 0     64D49M204D ASV1594   seq1
#> 1341    H       0   291     83.7      + 0 0     64D49M178D ASV1602   seq1
#> 1342    H       0   291     81.6      + 0 0     64D49M178D ASV1605   seq1
#> 1343    H       0   295     85.7      + 0 0 64D43M6D6M176D ASV1603   seq1
#> 1345    H       0   346     83.7      + 0 0     66D49M231D ASV1604   seq1
#> 1354    H       0   332     87.8      + 0 0     64D49M219D ASV1615   seq1
#> 1355    H       0   313     89.8      + 0 0     64D49M200D ASV1616   seq1
#> 1361    H       0   304     83.7      + 0 0     65D49M190D ASV1625   seq1
#> 1362    H       0   302     83.7      + 0 0     64D49M189D ASV1628   seq1
#> 1371    H       0   340     85.7      + 0 0 64D43M5D6M222D ASV1636   seq1
#> 1373    H       0   305     83.7      + 0 0     64D49M192D ASV1641   seq1
#> 1374    H       0   297     83.7      + 0 0 64D42M6D7M178D ASV1642   seq1
#> 1381    H       0   297     83.7      + 0 0     65D49M183D ASV1653   seq1
#> 1382    H       0   322     85.7      + 0 0     64D49M209D ASV1654   seq1
#> 1383    H       0   291     81.6      + 0 0     65D49M177D ASV1656   seq1
#> 1386    H       0   296     85.7      + 0 0 64D42M6D7M177D ASV1661   seq1
#> 1388    H       0   302     85.7      + 0 0     65D49M188D ASV1663   seq1
#> 1395    H       0   293     81.6      + 0 0     65D49M179D ASV1672   seq1
#> 1410    H       0   301     81.6      + 0 0     65D49M187D ASV1704   seq1

clean_pq(subset_taxa(data_fungi, res$identity != "*"))
#> Error in eval(e, x, parent.frame()): object 'res' not found
# }