Save phyloseq object in the form of multiple csv tables.

Usage

write_pq(
  physeq,
  path = NULL,
  rdata = FALSE,
  one_file_ASV = FALSE,
  write_sam_data = TRUE,
  ...
)

Arguments

physeq

(required) a phyloseq-class object.

path

a path to the folder to save the phyloseq object

rdata

(logical) does the phyloseq object is also saved in Rdata format?

one_file_ASV

(logical) if TRUE, combine all data in one file only

write_sam_data

(logical) does the samples data are add to the file. Only used if one_file_ASV is TRUE. Note that this result in a lot of NA values.

...

Other arguments passed to utils::write.csv() function.

Value

One to four csv tables (refseq.csv, otu_table.csv, tax_table.csv, sam_data.csv) and if present a phy_tree in Newick format

Examples

if (FALSE) {
write_pq(data_fungi, path = "phyloseq")
write_pq(data_fungi, path = "phyloseq", one_file_ASV = TRUE)
}