Compute and test local contributions to beta diversity (LCBD) of samples

A wrapper for the adespatial::beta.div() function in the case of physeq object.

Usage

LCBD_pq(physeq, p_adjust_method = "BH", ...)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

p_adjust_method

(chr, default "BH"): the method used to adjust p-value

...

Others arguments passed on to adespatial::beta.div() function

Value

An object of class beta.div see adespatial::beta.div() function for more information

See also

Author

Adrien Taudière This function is mainly a wrapper of the work of others. Please make a reference to adespatial::beta.div() if you use this function.

Examples

res <- LCBD_pq(data_fungi, nperm = 100)
str(res)
#> List of 8
#>  $ beta  : Named num [1:2] 175.33 0.953
#>   ..- attr(*, "names")= chr [1:2] "SStotal" "BDtotal"
#>  $ SCBD  : Named num [1:1420] 0.0379 0.0192 0.0148 0.0145 0.0128 ...
#>   ..- attr(*, "names")= chr [1:1420] "ASV2" "ASV6" "ASV7" "ASV8" ...
#>  $ LCBD  : Named num [1:185] 0.00535 0.00536 0.00531 0.00512 0.00514 ...
#>   ..- attr(*, "names")= chr [1:185] "A10-005-B_S188_MERGED.fastq.gz" "A10-005-H_S189_MERGED.fastq.gz" "A10-005-M_S190_MERGED.fastq.gz" "A12-007_S191_MERGED.fastq.gz" ...
#>  $ p.LCBD: Named num [1:185] 0.713 0.594 0.733 0.96 0.931 ...
#>   ..- attr(*, "names")= chr [1:185] "A10-005-B_S188_MERGED.fastq.gz" "A10-005-H_S189_MERGED.fastq.gz" "A10-005-M_S190_MERGED.fastq.gz" "A12-007_S191_MERGED.fastq.gz" ...
#>  $ p.adj : Named num [1:185] 0.995 0.995 0.995 0.995 0.995 ...
#>   ..- attr(*, "names")= chr [1:185] "A10-005-B_S188_MERGED.fastq.gz" "A10-005-H_S189_MERGED.fastq.gz" "A10-005-M_S190_MERGED.fastq.gz" "A12-007_S191_MERGED.fastq.gz" ...
#>  $ method: chr [1:2] "hellinger" NA
#>  $ note  : chr "Info -- This coefficient is Euclidean"
#>  $ D     : logi NA
#>  - attr(*, "class")= chr "beta.div"
length(res$LCBD)
#> [1] 185
length(res$SCBD)
#> [1] 1420
LCBD_pq(data_fungi, nperm = 100, method = "jaccard")
#> $beta
#>    SStotal    BDtotal 
#> 87.0121620  0.4728922 
#> 
#> $SCBD
#> [1] NA
#> 
#> $LCBD
#>    A10-005-B_S188_MERGED.fastq.gz    A10-005-H_S189_MERGED.fastq.gz 
#>                       0.005293759                       0.005476743 
#>    A10-005-M_S190_MERGED.fastq.gz      A12-007_S191_MERGED.fastq.gz 
#>                       0.005398885                       0.005269386 
#>      A12-007-B_S2_MERGED.fastq.gz        A15-004_S3_MERGED.fastq.gz 
#>                       0.005523163                       0.005156302 
#>         A8-005_S4_MERGED.fastq.gz         A9-012_S5_MERGED.fastq.gz 
#>                       0.005391397                       0.005350236 
#>    AB29-ABMX-H_S6_MERGED.fastq.gz       AC27-013_S7_MERGED.fastq.gz 
#>                       0.005594155                       0.005476597 
#>        AC29033_S8_MERGED.fastq.gz     AD26-005-B_S9_MERGED.fastq.gz 
#>                       0.005228302                       0.005352637 
#>    AD26-005-H_S10_MERGED.fastq.gz    AD26-005-M_S11_MERGED.fastq.gz 
#>                       0.005286048                       0.005329776 
#>   AD30-ABMX-M_S12_MERGED.fastq.gz    AD32-007-M_S13_MERGED.fastq.gz 
#>                       0.005506116                       0.005333049 
#>    ADABM30X-B_S14_MERGED.fastq.gz    ADABM30X-H_S15_MERGED.fastq.gz 
#>                       0.005493405                       0.005426811 
#>    ADABM30X-M_S16_MERGED.fastq.gz   AE30-ABM507_S17_MERGED.fastq.gz 
#>                       0.005593807                       0.005240489 
#>       B17-014_S18_MERGED.fastq.gz     B18-006-B_S19_MERGED.fastq.gz 
#>                       0.005094468                       0.005199642 
#>   BA16-036bis_S20_MERGED.fastq.gz    BA17-050-B_S21_MERGED.fastq.gz 
#>                       0.005602428                       0.005113083 
#>    BB19-006-H_S22_MERGED.fastq.gz     BB6-019-B_S23_MERGED.fastq.gz 
#>                       0.005315898                       0.005398123 
#>     BB6-019-H_S24_MERGED.fastq.gz     BB6-019-M_S25_MERGED.fastq.gz 
#>                       0.005399550                       0.005441915 
#>      BD14-021_S26_MERGED.fastq.gz     BE9-006-B_S27_MERGED.fastq.gz 
#>                       0.005214233                       0.005455816 
#>     BE9-006-H_S28_MERGED.fastq.gz     BE9-006-M_S29_MERGED.fastq.gz 
#>                       0.005547131                       0.005463533 
#>     BG7-010-B_S30_MERGED.fastq.gz     BG7-010-H_S31_MERGED.fastq.gz 
#>                       0.005122961                       0.005298004 
#>     BG7-010-M_S32_MERGED.fastq.gz       BH9-021_S33_MERGED.fastq.gz 
#>                       0.005155033                       0.005480963 
#>    BJ17-007-M_S34_MERGED.fastq.gz   BJ8-ABM-003_S35_MERGED.fastq.gz 
#>                       0.005240752                       0.005606386 
#>     BL7-006-B_S36_MERGED.fastq.gz     BL7-006-H_S37_MERGED.fastq.gz 
#>                       0.005128002                       0.005119007 
#>     BL7-006-M_S38_MERGED.fastq.gz      BN11-041_S39_MERGED.fastq.gz 
#>                       0.004998552                       0.005214421 
#>       BN9-002_S40_MERGED.fastq.gz       BO8-002_S41_MERGED.fastq.gz 
#>                       0.005693698                       0.005453127 
#>       BO8-005_S42_MERGED.fastq.gz    BP11-001-B_S43_MERGED.fastq.gz 
#>                       0.005748696                       0.005470716 
#>    BP11-001-H_S44_MERGED.fastq.gz    BP11-001-M_S45_MERGED.fastq.gz 
#>                       0.005263127                       0.005419313 
#>    BP12-025-B_S46_MERGED.fastq.gz      BP14-006_S47_MERGED.fastq.gz 
#>                       0.005554078                       0.005167507 
#>       BQ3-019_S48_MERGED.fastq.gz     BQ4-018-B_S49_MERGED.fastq.gz 
#>                       0.005605772                       0.005228773 
#>     BQ4-018-H_S50_MERGED.fastq.gz     BQ4-018-M_S51_MERGED.fastq.gz 
#>                       0.005486120                       0.005318799 
#>    BQ9ABM-002_S52_MERGED.fastq.gz       BR8-005_S53_MERGED.fastq.gz 
#>                       0.005683018                       0.005466321 
#>      BS14-006_S54_MERGED.fastq.gz      BT-006-M_S55_MERGED.fastq.gz 
#>                       0.005200842                       0.005244569 
#>       BT7-006_S56_MERGED.fastq.gz    BV11-002-B_S57_MERGED.fastq.gz 
#>                       0.005215359                       0.005380090 
#>    BV11-002-H_S58_MERGED.fastq.gz    BV11-002-M_S59_MERGED.fastq.gz 
#>                       0.005356740                       0.005494453 
#>       BW8-003_S60_MERGED.fastq.gz        C1-001_S61_MERGED.fastq.gz 
#>                       0.005142749                       0.005250146 
#>     C21-NV1-B_S62_MERGED.fastq.gz     C21-NV1-H_S63_MERGED.fastq.gz 
#>                       0.005498278                       0.005562212 
#>     C21-NV1-M_S64_MERGED.fastq.gz        C9-005_S65_MERGED.fastq.gz 
#>                       0.005668334                       0.005602974 
#>      CA12-024_S66_MERGED.fastq.gz       CA9-027_S67_MERGED.fastq.gz 
#>                       0.005698885                       0.005831616 
#>         CA9-X_S68_MERGED.fastq.gz     CB8-019-B_S69_MERGED.fastq.gz 
#>                       0.005381083                       0.005574658 
#>     CB8-019-H_S70_MERGED.fastq.gz     CB8-019-M_S71_MERGED.fastq.gz 
#>                       0.005525686                       0.005549116 
#>       CB9-013_S72_MERGED.fastq.gz       CC3-044_S73_MERGED.fastq.gz 
#>                       0.005399184                       0.005190772 
#>       CC8-003_S74_MERGED.fastq.gz       D17-011_S77_MERGED.fastq.gz 
#>                       0.005599282                       0.005333026 
#>     D18-003-B_S78_MERGED.fastq.gz     D18-003-H_S79_MERGED.fastq.gz 
#>                       0.005353657                       0.005465353 
#>     D18-003-M_S80_MERGED.fastq.gz    D22-NVABM1_S81_MERGED.fastq.gz 
#>                       0.005495151                       0.005288639 
#>     D61-010-B_S82_MERGED.fastq.gz      D9-027-B_S83_MERGED.fastq.gz 
#>                       0.005587602                       0.005291541 
#>      D9-027-H_S84_MERGED.fastq.gz      D9-027-M_S85_MERGED.fastq.gz 
#>                       0.005238435                       0.005347453 
#>   DBM-ABM-001_S86_MERGED.fastq.gz     DJ2-008-B_S87_MERGED.fastq.gz 
#>                       0.005229905                       0.005453967 
#>     DJ2-008-H_S88_MERGED.fastq.gz     DJ2-008-M_S89_MERGED.fastq.gz 
#>                       0.005556785                       0.005365977 
#>    DP4-ABM001_S90_MERGED.fastq.gz  DS1-ABM002-B_S91_MERGED.fastq.gz 
#>                       0.005663019                       0.005329258 
#>  DS1-ABM002-H_S92_MERGED.fastq.gz  DS1-ABM002-M_S93_MERGED.fastq.gz 
#>                       0.005521295                       0.005318572 
#>     DU3-045-B_S94_MERGED.fastq.gz       DW4-007_S95_MERGED.fastq.gz 
#>                       0.005296629                       0.005541464 
#>     DY5-004-B_S96_MERGED.fastq.gz     DY5-004-H_S97_MERGED.fastq.gz 
#>                       0.005262574                       0.005244919 
#>     DY5-004-M_S98_MERGED.fastq.gz   DZ6-ABM-001_S99_MERGED.fastq.gz 
#>                       0.005184916                       0.005515914 
#>     E9-009-B_S100_MERGED.fastq.gz     E9-009-H_S101_MERGED.fastq.gz 
#>                       0.005195336                       0.005096557 
#>     E9-009-M_S102_MERGED.fastq.gz  EA5-ABM-001_S103_MERGED.fastq.gz 
#>                       0.005131554                       0.005306707 
#>    EC2-013-B_S104_MERGED.fastq.gz   F6-ABM-001_S105_MERGED.fastq.gz 
#>                       0.005509238                       0.005477991 
#>     F7-015-M_S106_MERGED.fastq.gz    FOMES19-H_S108_MERGED.fastq.gz 
#>                       0.005630568                       0.005219386 
#>    FOMES19-M_S109_MERGED.fastq.gz    H10-018-M_S110_MERGED.fastq.gz 
#>                       0.005344141                       0.005732686 
#> H24-NVABM1-H_S111_MERGED.fastq.gz    J18-004-B_S114_MERGED.fastq.gz 
#>                       0.005353536                       0.005326498 
#>    J18-004-H_S115_MERGED.fastq.gz    J18-004-M_S116_MERGED.fastq.gz 
#>                       0.005667462                       0.005281526 
#>    K18-002-H_S117_MERGED.fastq.gz   K26-NVABM1_S118_MERGED.fastq.gz 
#>                       0.005416683                       0.005385941 
#>       L19X-B_S119_MERGED.fastq.gz       L19X-H_S120_MERGED.fastq.gz 
#>                       0.005637971                       0.005329344 
#>       L19X-M_S121_MERGED.fastq.gz    L23-002-B_S122_MERGED.fastq.gz 
#>                       0.005223836                       0.005120540 
#>    L23-002-H_S123_MERGED.fastq.gz    L23-002-M_S124_MERGED.fastq.gz 
#>                       0.005446937                       0.005486874 
#>      M22-001_S125_MERGED.fastq.gz       N19X-B_S126_MERGED.fastq.gz 
#>                       0.005638493                       0.005290225 
#>       N19X-H_S127_MERGED.fastq.gz       N19X-M_S128_MERGED.fastq.gz 
#>                       0.005141082                       0.005343979 
#>    N22-001-B_S129_MERGED.fastq.gz    N23-002-B_S130_MERGED.fastq.gz 
#>                       0.005736107                       0.005256235 
#>    N23-002-H_S131_MERGED.fastq.gz    N23-002-M_S132_MERGED.fastq.gz 
#>                       0.005462942                       0.005966039 
#>     N25-ABMX_S133_MERGED.fastq.gz   NVABM-0058_S134_MERGED.fastq.gz 
#>                       0.005494005                       0.005487254 
#> NVABM-0163-H_S135_MERGED.fastq.gz   NVABM-0397_S138_MERGED.fastq.gz 
#>                       0.005312855                       0.005362471 
#>    NVABM0216_S136_MERGED.fastq.gz  NVABM0244-M_S137_MERGED.fastq.gz 
#>                       0.005469529                       0.005547089 
#>      O20-X-B_S139_MERGED.fastq.gz      O20-X-H_S140_MERGED.fastq.gz 
#>                       0.005511046                       0.005211704 
#>      O20-X-M_S141_MERGED.fastq.gz    O21-007-B_S142_MERGED.fastq.gz 
#>                       0.005227456                       0.005499556 
#>    O21-007-H_S143_MERGED.fastq.gz    O21-007-M_S144_MERGED.fastq.gz 
#>                       0.005561784                       0.005794666 
#>    O24-003-B_S145_MERGED.fastq.gz    O24-003-H_S146_MERGED.fastq.gz 
#>                       0.005510677                       0.005588565 
#>    O24-003-M_S147_MERGED.fastq.gz    O26-004-B_S148_MERGED.fastq.gz 
#>                       0.005465284                       0.005539860 
#>    O26-004-H_S149_MERGED.fastq.gz    O26-004-M_S150_MERGED.fastq.gz 
#>                       0.005615870                       0.005835373 
#>      O27-012_S151_MERGED.fastq.gz     O9-005-B_S152_MERGED.fastq.gz 
#>                       0.005675974                       0.005288555 
#>    P19-023-M_S153_MERGED.fastq.gz    P27-015-M_S154_MERGED.fastq.gz 
#>                       0.005340713                       0.005611739 
#>   P27-ABM001_S155_MERGED.fastq.gz Q27-ABM003-B_S156_MERGED.fastq.gz 
#>                       0.005654546                       0.005446170 
#>     R25-ABMX_S157_MERGED.fastq.gz    R28-008-B_S158_MERGED.fastq.gz 
#>                       0.005240496                       0.005261409 
#>    R28-008-H_S159_MERGED.fastq.gz    R28-008-M_S160_MERGED.fastq.gz 
#>                       0.005250442                       0.005116651 
#>      T28-011_S161_MERGED.fastq.gz T28-ABM602-B_S162_MERGED.fastq.gz 
#>                       0.005547853                       0.005661540 
#>   U27-ABM002_S163_MERGED.fastq.gz     W25-ABMX_S164_MERGED.fastq.gz 
#>                       0.005216135                       0.005240407 
#>    W26-001-B_S165_MERGED.fastq.gz    W26-001-H_S166_MERGED.fastq.gz 
#>                       0.005109880                       0.005112376 
#>    W26-001-M_S167_MERGED.fastq.gz      W30-006_S168_MERGED.fastq.gz 
#>                       0.005293499                       0.005178635 
#>     W9-025-M_S169_MERGED.fastq.gz    X24-009-B_S170_MERGED.fastq.gz 
#>                       0.005602794                       0.005429243 
#>    X24-009-H_S171_MERGED.fastq.gz    X24-009-M_S172_MERGED.fastq.gz 
#>                       0.005689228                       0.005603163 
#>      X24-010_S173_MERGED.fastq.gz    X29-004-B_S174_MERGED.fastq.gz 
#>                       0.005667319                       0.005736969 
#>    X29-004-H_S175_MERGED.fastq.gz    X29-004-M_S176_MERGED.fastq.gz 
#>                       0.005137753                       0.005287792 
#> Y21-ABM484-H_S177_MERGED.fastq.gz    Y28-002-B_S178_MERGED.fastq.gz 
#>                       0.005440118                       0.005479090 
#>    Y28-002-H_S179_MERGED.fastq.gz    Y28-002-M_S180_MERGED.fastq.gz 
#>                       0.005415747                       0.005539610 
#>    Y29-007-B_S181_MERGED.fastq.gz    Y29-007-H_S182_MERGED.fastq.gz 
#>                       0.005180327                       0.005182281 
#>    Y29-007-M_S183_MERGED.fastq.gz Y31-ABM484-B_S184_MERGED.fastq.gz 
#>                       0.005445237                       0.005494207 
#>    Z29-001-H_S185_MERGED.fastq.gz      Z30-002_S186_MERGED.fastq.gz 
#>                       0.005405564                       0.005319007 
#> Z30-ABM560-M_S187_MERGED.fastq.gz 
#>                       0.005362959 
#> 
#> $p.LCBD
#>    A10-005-B_S188_MERGED.fastq.gz    A10-005-H_S189_MERGED.fastq.gz 
#>                        0.90099010                        0.14851485 
#>    A10-005-M_S190_MERGED.fastq.gz      A12-007_S191_MERGED.fastq.gz 
#>                        0.46534653                        0.97029703 
#>      A12-007-B_S2_MERGED.fastq.gz        A15-004_S3_MERGED.fastq.gz 
#>                        0.09900990                        1.00000000 
#>         A8-005_S4_MERGED.fastq.gz         A9-012_S5_MERGED.fastq.gz 
#>                        0.46534653                        0.75247525 
#>    AB29-ABMX-H_S6_MERGED.fastq.gz       AC27-013_S7_MERGED.fastq.gz 
#>                        0.00990099                        0.18811881 
#>        AC29033_S8_MERGED.fastq.gz     AD26-005-B_S9_MERGED.fastq.gz 
#>                        0.97029703                        0.81188119 
#>    AD26-005-H_S10_MERGED.fastq.gz    AD26-005-M_S11_MERGED.fastq.gz 
#>                        0.98019802                        0.89108911 
#>   AD30-ABMX-M_S12_MERGED.fastq.gz    AD32-007-M_S13_MERGED.fastq.gz 
#>                        0.06930693                        0.86138614 
#>    ADABM30X-B_S14_MERGED.fastq.gz    ADABM30X-H_S15_MERGED.fastq.gz 
#>                        0.07920792                        0.39603960 
#>    ADABM30X-M_S16_MERGED.fastq.gz   AE30-ABM507_S17_MERGED.fastq.gz 
#>                        0.00990099                        1.00000000 
#>       B17-014_S18_MERGED.fastq.gz     B18-006-B_S19_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>   BA16-036bis_S20_MERGED.fastq.gz    BA17-050-B_S21_MERGED.fastq.gz 
#>                        0.00990099                        1.00000000 
#>    BB19-006-H_S22_MERGED.fastq.gz     BB6-019-B_S23_MERGED.fastq.gz 
#>                        0.92079208                        0.53465347 
#>     BB6-019-H_S24_MERGED.fastq.gz     BB6-019-M_S25_MERGED.fastq.gz 
#>                        0.44554455                        0.27722772 
#>      BD14-021_S26_MERGED.fastq.gz     BE9-006-B_S27_MERGED.fastq.gz 
#>                        1.00000000                        0.18811881 
#>     BE9-006-H_S28_MERGED.fastq.gz     BE9-006-M_S29_MERGED.fastq.gz 
#>                        0.02970297                        0.20792079 
#>     BG7-010-B_S30_MERGED.fastq.gz     BG7-010-H_S31_MERGED.fastq.gz 
#>                        1.00000000                        0.93069307 
#>     BG7-010-M_S32_MERGED.fastq.gz       BH9-021_S33_MERGED.fastq.gz 
#>                        1.00000000                        0.12871287 
#>    BJ17-007-M_S34_MERGED.fastq.gz   BJ8-ABM-003_S35_MERGED.fastq.gz 
#>                        1.00000000                        0.00990099 
#>     BL7-006-B_S36_MERGED.fastq.gz     BL7-006-H_S37_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     BL7-006-M_S38_MERGED.fastq.gz      BN11-041_S39_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>       BN9-002_S40_MERGED.fastq.gz       BO8-002_S41_MERGED.fastq.gz 
#>                        0.00990099                        0.26732673 
#>       BO8-005_S42_MERGED.fastq.gz    BP11-001-B_S43_MERGED.fastq.gz 
#>                        0.00990099                        0.18811881 
#>    BP11-001-H_S44_MERGED.fastq.gz    BP11-001-M_S45_MERGED.fastq.gz 
#>                        0.98019802                        0.39603960 
#>    BP12-025-B_S46_MERGED.fastq.gz      BP14-006_S47_MERGED.fastq.gz 
#>                        0.02970297                        1.00000000 
#>       BQ3-019_S48_MERGED.fastq.gz     BQ4-018-B_S49_MERGED.fastq.gz 
#>                        0.00990099                        0.99009901 
#>     BQ4-018-H_S50_MERGED.fastq.gz     BQ4-018-M_S51_MERGED.fastq.gz 
#>                        0.09900990                        0.86138614 
#>    BQ9ABM-002_S52_MERGED.fastq.gz       BR8-005_S53_MERGED.fastq.gz 
#>                        0.00990099                        0.22772277 
#>      BS14-006_S54_MERGED.fastq.gz      BT-006-M_S55_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>       BT7-006_S56_MERGED.fastq.gz    BV11-002-B_S57_MERGED.fastq.gz 
#>                        0.99009901                        0.63366337 
#>    BV11-002-H_S58_MERGED.fastq.gz    BV11-002-M_S59_MERGED.fastq.gz 
#>                        0.78217822                        0.11881188 
#>       BW8-003_S60_MERGED.fastq.gz        C1-001_S61_MERGED.fastq.gz 
#>                        1.00000000                        0.98019802 
#>     C21-NV1-B_S62_MERGED.fastq.gz     C21-NV1-H_S63_MERGED.fastq.gz 
#>                        0.13861386                        0.00990099 
#>     C21-NV1-M_S64_MERGED.fastq.gz        C9-005_S65_MERGED.fastq.gz 
#>                        0.00990099                        0.00990099 
#>      CA12-024_S66_MERGED.fastq.gz       CA9-027_S67_MERGED.fastq.gz 
#>                        0.00990099                        0.00990099 
#>         CA9-X_S68_MERGED.fastq.gz     CB8-019-B_S69_MERGED.fastq.gz 
#>                        0.63366337                        0.02970297 
#>     CB8-019-H_S70_MERGED.fastq.gz     CB8-019-M_S71_MERGED.fastq.gz 
#>                        0.05940594                        0.03960396 
#>       CB9-013_S72_MERGED.fastq.gz       CC3-044_S73_MERGED.fastq.gz 
#>                        0.55445545                        1.00000000 
#>       CC8-003_S74_MERGED.fastq.gz       D17-011_S77_MERGED.fastq.gz 
#>                        0.00990099                        0.85148515 
#>     D18-003-B_S78_MERGED.fastq.gz     D18-003-H_S79_MERGED.fastq.gz 
#>                        0.79207921                        0.26732673 
#>     D18-003-M_S80_MERGED.fastq.gz    D22-NVABM1_S81_MERGED.fastq.gz 
#>                        0.14851485                        0.96039604 
#>     D61-010-B_S82_MERGED.fastq.gz      D9-027-B_S83_MERGED.fastq.gz 
#>                        0.00990099                        0.96039604 
#>      D9-027-H_S84_MERGED.fastq.gz      D9-027-M_S85_MERGED.fastq.gz 
#>                        1.00000000                        0.78217822 
#>   DBM-ABM-001_S86_MERGED.fastq.gz     DJ2-008-B_S87_MERGED.fastq.gz 
#>                        0.99009901                        0.28712871 
#>     DJ2-008-H_S88_MERGED.fastq.gz     DJ2-008-M_S89_MERGED.fastq.gz 
#>                        0.02970297                        0.76237624 
#>    DP4-ABM001_S90_MERGED.fastq.gz  DS1-ABM002-B_S91_MERGED.fastq.gz 
#>                        0.00990099                        0.91089109 
#>  DS1-ABM002-H_S92_MERGED.fastq.gz  DS1-ABM002-M_S93_MERGED.fastq.gz 
#>                        0.03960396                        0.90099010 
#>     DU3-045-B_S94_MERGED.fastq.gz       DW4-007_S95_MERGED.fastq.gz 
#>                        0.96039604                        0.02970297 
#>     DY5-004-B_S96_MERGED.fastq.gz     DY5-004-H_S97_MERGED.fastq.gz 
#>                        0.99009901                        1.00000000 
#>     DY5-004-M_S98_MERGED.fastq.gz   DZ6-ABM-001_S99_MERGED.fastq.gz 
#>                        1.00000000                        0.04950495 
#>     E9-009-B_S100_MERGED.fastq.gz     E9-009-H_S101_MERGED.fastq.gz 
#>                        0.99009901                        1.00000000 
#>     E9-009-M_S102_MERGED.fastq.gz  EA5-ABM-001_S103_MERGED.fastq.gz 
#>                        1.00000000                        0.87128713 
#>    EC2-013-B_S104_MERGED.fastq.gz   F6-ABM-001_S105_MERGED.fastq.gz 
#>                        0.06930693                        0.14851485 
#>     F7-015-M_S106_MERGED.fastq.gz    FOMES19-H_S108_MERGED.fastq.gz 
#>                        0.00990099                        1.00000000 
#>    FOMES19-M_S109_MERGED.fastq.gz    H10-018-M_S110_MERGED.fastq.gz 
#>                        0.85148515                        0.00990099 
#> H24-NVABM1-H_S111_MERGED.fastq.gz    J18-004-B_S114_MERGED.fastq.gz 
#>                        0.72277228                        0.93069307 
#>    J18-004-H_S115_MERGED.fastq.gz    J18-004-M_S116_MERGED.fastq.gz 
#>                        0.00990099                        0.97029703 
#>    K18-002-H_S117_MERGED.fastq.gz   K26-NVABM1_S118_MERGED.fastq.gz 
#>                        0.47524752                        0.60396040 
#>       L19X-B_S119_MERGED.fastq.gz       L19X-H_S120_MERGED.fastq.gz 
#>                        0.00990099                        0.90099010 
#>       L19X-M_S121_MERGED.fastq.gz    L23-002-B_S122_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    L23-002-H_S123_MERGED.fastq.gz    L23-002-M_S124_MERGED.fastq.gz 
#>                        0.34653465                        0.13861386 
#>      M22-001_S125_MERGED.fastq.gz       N19X-B_S126_MERGED.fastq.gz 
#>                        0.00990099                        0.96039604 
#>       N19X-H_S127_MERGED.fastq.gz       N19X-M_S128_MERGED.fastq.gz 
#>                        1.00000000                        0.79207921 
#>    N22-001-B_S129_MERGED.fastq.gz    N23-002-B_S130_MERGED.fastq.gz 
#>                        0.00990099                        0.98019802 
#>    N23-002-H_S131_MERGED.fastq.gz    N23-002-M_S132_MERGED.fastq.gz 
#>                        0.14851485                        0.00990099 
#>     N25-ABMX_S133_MERGED.fastq.gz   NVABM-0058_S134_MERGED.fastq.gz 
#>                        0.08910891                        0.13861386 
#> NVABM-0163-H_S135_MERGED.fastq.gz   NVABM-0397_S138_MERGED.fastq.gz 
#>                        0.91089109                        0.75247525 
#>    NVABM0216_S136_MERGED.fastq.gz  NVABM0244-M_S137_MERGED.fastq.gz 
#>                        0.24752475                        0.02970297 
#>      O20-X-B_S139_MERGED.fastq.gz      O20-X-H_S140_MERGED.fastq.gz 
#>                        0.07920792                        1.00000000 
#>      O20-X-M_S141_MERGED.fastq.gz    O21-007-B_S142_MERGED.fastq.gz 
#>                        0.99009901                        0.09900990 
#>    O21-007-H_S143_MERGED.fastq.gz    O21-007-M_S144_MERGED.fastq.gz 
#>                        0.03960396                        0.00990099 
#>    O24-003-B_S145_MERGED.fastq.gz    O24-003-H_S146_MERGED.fastq.gz 
#>                        0.06930693                        0.01980198 
#>    O24-003-M_S147_MERGED.fastq.gz    O26-004-B_S148_MERGED.fastq.gz 
#>                        0.21782178                        0.04950495 
#>    O26-004-H_S149_MERGED.fastq.gz    O26-004-M_S150_MERGED.fastq.gz 
#>                        0.00990099                        0.00990099 
#>      O27-012_S151_MERGED.fastq.gz     O9-005-B_S152_MERGED.fastq.gz 
#>                        0.00990099                        0.96039604 
#>    P19-023-M_S153_MERGED.fastq.gz    P27-015-M_S154_MERGED.fastq.gz 
#>                        0.84158416                        0.00990099 
#>   P27-ABM001_S155_MERGED.fastq.gz Q27-ABM003-B_S156_MERGED.fastq.gz 
#>                        0.00990099                        0.26732673 
#>     R25-ABMX_S157_MERGED.fastq.gz    R28-008-B_S158_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    R28-008-H_S159_MERGED.fastq.gz    R28-008-M_S160_MERGED.fastq.gz 
#>                        0.98019802                        1.00000000 
#>      T28-011_S161_MERGED.fastq.gz T28-ABM602-B_S162_MERGED.fastq.gz 
#>                        0.00990099                        0.00990099 
#>   U27-ABM002_S163_MERGED.fastq.gz     W25-ABMX_S164_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    W26-001-B_S165_MERGED.fastq.gz    W26-001-H_S166_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    W26-001-M_S167_MERGED.fastq.gz      W30-006_S168_MERGED.fastq.gz 
#>                        0.98019802                        1.00000000 
#>     W9-025-M_S169_MERGED.fastq.gz    X24-009-B_S170_MERGED.fastq.gz 
#>                        0.00990099                        0.35643564 
#>    X24-009-H_S171_MERGED.fastq.gz    X24-009-M_S172_MERGED.fastq.gz 
#>                        0.00990099                        0.00990099 
#>      X24-010_S173_MERGED.fastq.gz    X29-004-B_S174_MERGED.fastq.gz 
#>                        0.00990099                        0.00990099 
#>    X29-004-H_S175_MERGED.fastq.gz    X29-004-M_S176_MERGED.fastq.gz 
#>                        0.99009901                        0.98019802 
#> Y21-ABM484-H_S177_MERGED.fastq.gz    Y28-002-B_S178_MERGED.fastq.gz 
#>                        0.27722772                        0.17821782 
#>    Y28-002-H_S179_MERGED.fastq.gz    Y28-002-M_S180_MERGED.fastq.gz 
#>                        0.40594059                        0.01980198 
#>    Y29-007-B_S181_MERGED.fastq.gz    Y29-007-H_S182_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    Y29-007-M_S183_MERGED.fastq.gz Y31-ABM484-B_S184_MERGED.fastq.gz 
#>                        0.23762376                        0.11881188 
#>    Z29-001-H_S185_MERGED.fastq.gz      Z30-002_S186_MERGED.fastq.gz 
#>                        0.58415842                        0.92079208 
#> Z30-ABM560-M_S187_MERGED.fastq.gz 
#>                        0.73267327 
#> 
#> $p.adj
#>    A10-005-B_S188_MERGED.fastq.gz    A10-005-H_S189_MERGED.fastq.gz 
#>                        1.00000000                        0.39819199 
#>    A10-005-M_S190_MERGED.fastq.gz      A12-007_S191_MERGED.fastq.gz 
#>                        0.93575118                        1.00000000 
#>      A12-007-B_S2_MERGED.fastq.gz        A15-004_S3_MERGED.fastq.gz 
#>                        0.31045477                        1.00000000 
#>         A8-005_S4_MERGED.fastq.gz         A9-012_S5_MERGED.fastq.gz 
#>                        0.93575118                        1.00000000 
#>    AB29-ABMX-H_S6_MERGED.fastq.gz       AC27-013_S7_MERGED.fastq.gz 
#>                        0.05088009                        0.47673945 
#>        AC29033_S8_MERGED.fastq.gz     AD26-005-B_S9_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    AD26-005-H_S10_MERGED.fastq.gz    AD26-005-M_S11_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>   AD30-ABMX-M_S12_MERGED.fastq.gz    AD32-007-M_S13_MERGED.fastq.gz 
#>                        0.24192042                        1.00000000 
#>    ADABM30X-B_S14_MERGED.fastq.gz    ADABM30X-H_S15_MERGED.fastq.gz 
#>                        0.26642664                        0.83258326 
#>    ADABM30X-M_S16_MERGED.fastq.gz   AE30-ABM507_S17_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>       B17-014_S18_MERGED.fastq.gz     B18-006-B_S19_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>   BA16-036bis_S20_MERGED.fastq.gz    BA17-050-B_S21_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>    BB19-006-H_S22_MERGED.fastq.gz     BB6-019-B_S23_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     BB6-019-H_S24_MERGED.fastq.gz     BB6-019-M_S25_MERGED.fastq.gz 
#>                        0.91584158                        0.61791721 
#>      BD14-021_S26_MERGED.fastq.gz     BE9-006-B_S27_MERGED.fastq.gz 
#>                        1.00000000                        0.47673945 
#>     BE9-006-H_S28_MERGED.fastq.gz     BE9-006-M_S29_MERGED.fastq.gz 
#>                        0.12488749                        0.51980198 
#>     BG7-010-B_S30_MERGED.fastq.gz     BG7-010-H_S31_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     BG7-010-M_S32_MERGED.fastq.gz       BH9-021_S33_MERGED.fastq.gz 
#>                        1.00000000                        0.38406260 
#>    BJ17-007-M_S34_MERGED.fastq.gz   BJ8-ABM-003_S35_MERGED.fastq.gz 
#>                        1.00000000                        0.05088009 
#>     BL7-006-B_S36_MERGED.fastq.gz     BL7-006-H_S37_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     BL7-006-M_S38_MERGED.fastq.gz      BN11-041_S39_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>       BN9-002_S40_MERGED.fastq.gz       BO8-002_S41_MERGED.fastq.gz 
#>                        0.05088009                        0.61056106 
#>       BO8-005_S42_MERGED.fastq.gz    BP11-001-B_S43_MERGED.fastq.gz 
#>                        0.05088009                        0.47673945 
#>    BP11-001-H_S44_MERGED.fastq.gz    BP11-001-M_S45_MERGED.fastq.gz 
#>                        1.00000000                        0.83258326 
#>    BP12-025-B_S46_MERGED.fastq.gz      BP14-006_S47_MERGED.fastq.gz 
#>                        0.12488749                        1.00000000 
#>       BQ3-019_S48_MERGED.fastq.gz     BQ4-018-B_S49_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>     BQ4-018-H_S50_MERGED.fastq.gz     BQ4-018-M_S51_MERGED.fastq.gz 
#>                        0.31045477                        1.00000000 
#>    BQ9ABM-002_S52_MERGED.fastq.gz       BR8-005_S53_MERGED.fastq.gz 
#>                        0.05088009                        0.55432517 
#>      BS14-006_S54_MERGED.fastq.gz      BT-006-M_S55_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>       BT7-006_S56_MERGED.fastq.gz    BV11-002-B_S57_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    BV11-002-H_S58_MERGED.fastq.gz    BV11-002-M_S59_MERGED.fastq.gz 
#>                        1.00000000                        0.36033112 
#>       BW8-003_S60_MERGED.fastq.gz        C1-001_S61_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     C21-NV1-B_S62_MERGED.fastq.gz     C21-NV1-H_S63_MERGED.fastq.gz 
#>                        0.39451637                        0.05088009 
#>     C21-NV1-M_S64_MERGED.fastq.gz        C9-005_S65_MERGED.fastq.gz 
#>                        0.05088009                        0.05088009 
#>      CA12-024_S66_MERGED.fastq.gz       CA9-027_S67_MERGED.fastq.gz 
#>                        0.05088009                        0.05088009 
#>         CA9-X_S68_MERGED.fastq.gz     CB8-019-B_S69_MERGED.fastq.gz 
#>                        1.00000000                        0.12488749 
#>     CB8-019-H_S70_MERGED.fastq.gz     CB8-019-M_S71_MERGED.fastq.gz 
#>                        0.21980198                        0.15588793 
#>       CB9-013_S72_MERGED.fastq.gz       CC3-044_S73_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>       CC8-003_S74_MERGED.fastq.gz       D17-011_S77_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>     D18-003-B_S78_MERGED.fastq.gz     D18-003-H_S79_MERGED.fastq.gz 
#>                        1.00000000                        0.61056106 
#>     D18-003-M_S80_MERGED.fastq.gz    D22-NVABM1_S81_MERGED.fastq.gz 
#>                        0.39819199                        1.00000000 
#>     D61-010-B_S82_MERGED.fastq.gz      D9-027-B_S83_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>      D9-027-H_S84_MERGED.fastq.gz      D9-027-M_S85_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>   DBM-ABM-001_S86_MERGED.fastq.gz     DJ2-008-B_S87_MERGED.fastq.gz 
#>                        1.00000000                        0.63236681 
#>     DJ2-008-H_S88_MERGED.fastq.gz     DJ2-008-M_S89_MERGED.fastq.gz 
#>                        0.12488749                        1.00000000 
#>    DP4-ABM001_S90_MERGED.fastq.gz  DS1-ABM002-B_S91_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>  DS1-ABM002-H_S92_MERGED.fastq.gz  DS1-ABM002-M_S93_MERGED.fastq.gz 
#>                        0.15588793                        1.00000000 
#>     DU3-045-B_S94_MERGED.fastq.gz       DW4-007_S95_MERGED.fastq.gz 
#>                        1.00000000                        0.12488749 
#>     DY5-004-B_S96_MERGED.fastq.gz     DY5-004-H_S97_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     DY5-004-M_S98_MERGED.fastq.gz   DZ6-ABM-001_S99_MERGED.fastq.gz 
#>                        1.00000000                        0.18690645 
#>     E9-009-B_S100_MERGED.fastq.gz     E9-009-H_S101_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     E9-009-M_S102_MERGED.fastq.gz  EA5-ABM-001_S103_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    EC2-013-B_S104_MERGED.fastq.gz   F6-ABM-001_S105_MERGED.fastq.gz 
#>                        0.24192042                        0.39819199 
#>     F7-015-M_S106_MERGED.fastq.gz    FOMES19-H_S108_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>    FOMES19-M_S109_MERGED.fastq.gz    H10-018-M_S110_MERGED.fastq.gz 
#>                        1.00000000                        0.05088009 
#> H24-NVABM1-H_S111_MERGED.fastq.gz    J18-004-B_S114_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    J18-004-H_S115_MERGED.fastq.gz    J18-004-M_S116_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>    K18-002-H_S117_MERGED.fastq.gz   K26-NVABM1_S118_MERGED.fastq.gz 
#>                        0.94538486                        1.00000000 
#>       L19X-B_S119_MERGED.fastq.gz       L19X-H_S120_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>       L19X-M_S121_MERGED.fastq.gz    L23-002-B_S122_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    L23-002-H_S123_MERGED.fastq.gz    L23-002-M_S124_MERGED.fastq.gz 
#>                        0.75422248                        0.39451637 
#>      M22-001_S125_MERGED.fastq.gz       N19X-B_S126_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>       N19X-H_S127_MERGED.fastq.gz       N19X-M_S128_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    N22-001-B_S129_MERGED.fastq.gz    N23-002-B_S130_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>    N23-002-H_S131_MERGED.fastq.gz    N23-002-M_S132_MERGED.fastq.gz 
#>                        0.39819199                        0.05088009 
#>     N25-ABMX_S133_MERGED.fastq.gz   NVABM-0058_S134_MERGED.fastq.gz 
#>                        0.29437765                        0.39451637 
#> NVABM-0163-H_S135_MERGED.fastq.gz   NVABM-0397_S138_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    NVABM0216_S136_MERGED.fastq.gz  NVABM0244-M_S137_MERGED.fastq.gz 
#>                        0.58707794                        0.12488749 
#>      O20-X-B_S139_MERGED.fastq.gz      O20-X-H_S140_MERGED.fastq.gz 
#>                        0.26642664                        1.00000000 
#>      O20-X-M_S141_MERGED.fastq.gz    O21-007-B_S142_MERGED.fastq.gz 
#>                        1.00000000                        0.31045477 
#>    O21-007-H_S143_MERGED.fastq.gz    O21-007-M_S144_MERGED.fastq.gz 
#>                        0.15588793                        0.05088009 
#>    O24-003-B_S145_MERGED.fastq.gz    O24-003-H_S146_MERGED.fastq.gz 
#>                        0.24192042                        0.09640438 
#>    O24-003-M_S147_MERGED.fastq.gz    O26-004-B_S148_MERGED.fastq.gz 
#>                        0.53729373                        0.18690645 
#>    O26-004-H_S149_MERGED.fastq.gz    O26-004-M_S150_MERGED.fastq.gz 
#>                        0.05088009                        0.05088009 
#>      O27-012_S151_MERGED.fastq.gz     O9-005-B_S152_MERGED.fastq.gz 
#>                        0.05088009                        1.00000000 
#>    P19-023-M_S153_MERGED.fastq.gz    P27-015-M_S154_MERGED.fastq.gz 
#>                        1.00000000                        0.05088009 
#>   P27-ABM001_S155_MERGED.fastq.gz Q27-ABM003-B_S156_MERGED.fastq.gz 
#>                        0.05088009                        0.61056106 
#>     R25-ABMX_S157_MERGED.fastq.gz    R28-008-B_S158_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    R28-008-H_S159_MERGED.fastq.gz    R28-008-M_S160_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>      T28-011_S161_MERGED.fastq.gz T28-ABM602-B_S162_MERGED.fastq.gz 
#>                        0.05088009                        0.05088009 
#>   U27-ABM002_S163_MERGED.fastq.gz     W25-ABMX_S164_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    W26-001-B_S165_MERGED.fastq.gz    W26-001-H_S166_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    W26-001-M_S167_MERGED.fastq.gz      W30-006_S168_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>     W9-025-M_S169_MERGED.fastq.gz    X24-009-B_S170_MERGED.fastq.gz 
#>                        0.05088009                        0.76675109 
#>    X24-009-H_S171_MERGED.fastq.gz    X24-009-M_S172_MERGED.fastq.gz 
#>                        0.05088009                        0.05088009 
#>      X24-010_S173_MERGED.fastq.gz    X29-004-B_S174_MERGED.fastq.gz 
#>                        0.05088009                        0.05088009 
#>    X29-004-H_S175_MERGED.fastq.gz    X29-004-M_S176_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#> Y21-ABM484-H_S177_MERGED.fastq.gz    Y28-002-B_S178_MERGED.fastq.gz 
#>                        0.61791721                        0.47100424 
#>    Y28-002-H_S179_MERGED.fastq.gz    Y28-002-M_S180_MERGED.fastq.gz 
#>                        0.84380910                        0.09640438 
#>    Y29-007-B_S181_MERGED.fastq.gz    Y29-007-H_S182_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#>    Y29-007-M_S183_MERGED.fastq.gz Y31-ABM484-B_S184_MERGED.fastq.gz 
#>                        0.57091423                        0.36033112 
#>    Z29-001-H_S185_MERGED.fastq.gz      Z30-002_S186_MERGED.fastq.gz 
#>                        1.00000000                        1.00000000 
#> Z30-ABM560-M_S187_MERGED.fastq.gz 
#>                        1.00000000 
#> 
#> $method
#> [1] "jaccard"      "sqrt.D=FALSE"
#> 
#> $note
#> [1] "Info -- D is Euclidean because beta.div outputs D[jk] = sqrt(1-S[jk])"
#> [2] "For this D functions, use beta.div with option sqrt.D=FALSE"          
#> 
#> $D
#> [1] NA
#> 
#> attr(,"class")
#> [1] "beta.div"