Compute and test local contributions to beta diversity (LCBD) of samples

A wrapper for the adespatial::beta.div() function in the case of physeq object.

Usage

LCBD_pq(physeq, p_adjust_method = "BH", ...)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

p_adjust_method

(chr, default "BH"): the method used to adjust p-value

...

Others arguments passed on to adespatial::beta.div() function

Value

An object of class beta.div see adespatial::beta.div() function for more information

See also

Author

Adrien Taudière This function is mainly a wrapper of the work of others. Please make a reference to adespatial::beta.div() if you use this function.

Examples

res <- LCBD_pq(data_fungi_sp_known, nperm = 5)
str(res)
#> List of 8
#>  $ beta  : Named num [1:2] 173.647 0.949
#>   ..- attr(*, "names")= chr [1:2] "SStotal" "BDtotal"
#>  $ SCBD  : Named num [1:651] 0.0533 0.0307 0.0206 0.0123 0.0122 ...
#>   ..- attr(*, "names")= chr [1:651] "ASV2" "ASV8" "ASV12" "ASV18" ...
#>  $ LCBD  : Named num [1:184] 0.00522 0.00559 0.00545 0.00516 0.00488 ...
#>   ..- attr(*, "names")= chr [1:184] "A10-005-B_S188_MERGED.fastq.gz" "A10-005-H_S189_MERGED.fastq.gz" "A10-005-M_S190_MERGED.fastq.gz" "A12-007_S191_MERGED.fastq.gz" ...
#>  $ p.LCBD: Named num [1:184] 1 0.5 0.667 1 1 ...
#>   ..- attr(*, "names")= chr [1:184] "A10-005-B_S188_MERGED.fastq.gz" "A10-005-H_S189_MERGED.fastq.gz" "A10-005-M_S190_MERGED.fastq.gz" "A12-007_S191_MERGED.fastq.gz" ...
#>  $ p.adj : Named num [1:184] 1 0.948 1 1 1 ...
#>   ..- attr(*, "names")= chr [1:184] "A10-005-B_S188_MERGED.fastq.gz" "A10-005-H_S189_MERGED.fastq.gz" "A10-005-M_S190_MERGED.fastq.gz" "A12-007_S191_MERGED.fastq.gz" ...
#>  $ method: chr [1:2] "hellinger" NA
#>  $ note  : chr "Info -- This coefficient is Euclidean"
#>  $ D     : logi NA
#>  - attr(*, "class")= chr "beta.div"
length(res$LCBD)
#> [1] 184
length(res$SCBD)
#> [1] 651
# \donttest{
LCBD_pq(data_fungi_sp_known, nperm = 5, method = "jaccard")
#> $beta
#>    SStotal    BDtotal 
#> 85.9986349  0.4699379 
#> 
#> $SCBD
#> [1] NA
#> 
#> $LCBD
#>    A10-005-B_S188_MERGED.fastq.gz    A10-005-H_S189_MERGED.fastq.gz 
#>                       0.005431337                       0.005585492 
#>    A10-005-M_S190_MERGED.fastq.gz      A12-007_S191_MERGED.fastq.gz 
#>                       0.005478440                       0.005475019 
#>      A12-007-B_S2_MERGED.fastq.gz        A15-004_S3_MERGED.fastq.gz 
#>                       0.005495528                       0.005107525 
#>         A8-005_S4_MERGED.fastq.gz         A9-012_S5_MERGED.fastq.gz 
#>                       0.005393136                       0.005158405 
#>    AB29-ABMX-H_S6_MERGED.fastq.gz       AC27-013_S7_MERGED.fastq.gz 
#>                       0.005598452                       0.005362440 
#>        AC29033_S8_MERGED.fastq.gz     AD26-005-B_S9_MERGED.fastq.gz 
#>                       0.005435153                       0.005590336 
#>    AD26-005-H_S10_MERGED.fastq.gz    AD26-005-M_S11_MERGED.fastq.gz 
#>                       0.005541210                       0.005601384 
#>   AD30-ABMX-M_S12_MERGED.fastq.gz    AD32-007-M_S13_MERGED.fastq.gz 
#>                       0.005500683                       0.005415709 
#>    ADABM30X-B_S14_MERGED.fastq.gz    ADABM30X-H_S15_MERGED.fastq.gz 
#>                       0.005521256                       0.005515426 
#>    ADABM30X-M_S16_MERGED.fastq.gz   AE30-ABM507_S17_MERGED.fastq.gz 
#>                       0.005664555                       0.005417710 
#>       B17-014_S18_MERGED.fastq.gz     B18-006-B_S19_MERGED.fastq.gz 
#>                       0.005223051                       0.005124279 
#>   BA16-036bis_S20_MERGED.fastq.gz    BA17-050-B_S21_MERGED.fastq.gz 
#>                       0.005596074                       0.005383404 
#>    BB19-006-H_S22_MERGED.fastq.gz     BB6-019-B_S23_MERGED.fastq.gz 
#>                       0.005479949                       0.005473285 
#>     BB6-019-H_S24_MERGED.fastq.gz     BB6-019-M_S25_MERGED.fastq.gz 
#>                       0.005516622                       0.005405949 
#>      BD14-021_S26_MERGED.fastq.gz     BE9-006-B_S27_MERGED.fastq.gz 
#>                       0.005391203                       0.005416875 
#>     BE9-006-H_S28_MERGED.fastq.gz     BE9-006-M_S29_MERGED.fastq.gz 
#>                       0.005388380                       0.005428383 
#>     BG7-010-B_S30_MERGED.fastq.gz     BG7-010-H_S31_MERGED.fastq.gz 
#>                       0.004971520                       0.005251272 
#>     BG7-010-M_S32_MERGED.fastq.gz       BH9-021_S33_MERGED.fastq.gz 
#>                       0.004998273                       0.005289260 
#>    BJ17-007-M_S34_MERGED.fastq.gz   BJ8-ABM-003_S35_MERGED.fastq.gz 
#>                       0.005409853                       0.005612220 
#>     BL7-006-B_S36_MERGED.fastq.gz     BL7-006-H_S37_MERGED.fastq.gz 
#>                       0.005027453                       0.004991713 
#>     BL7-006-M_S38_MERGED.fastq.gz      BN11-041_S39_MERGED.fastq.gz 
#>                       0.004807057                       0.005462552 
#>       BN9-002_S40_MERGED.fastq.gz       BO8-002_S41_MERGED.fastq.gz 
#>                       0.005701409                       0.005552082 
#>       BO8-005_S42_MERGED.fastq.gz    BP11-001-B_S43_MERGED.fastq.gz 
#>                       0.005608939                       0.005528332 
#>    BP11-001-H_S44_MERGED.fastq.gz    BP11-001-M_S45_MERGED.fastq.gz 
#>                       0.005370893                       0.005529987 
#>    BP12-025-B_S46_MERGED.fastq.gz      BP14-006_S47_MERGED.fastq.gz 
#>                       0.005591550                       0.005358656 
#>       BQ3-019_S48_MERGED.fastq.gz     BQ4-018-B_S49_MERGED.fastq.gz 
#>                       0.005701273                       0.005389513 
#>     BQ4-018-H_S50_MERGED.fastq.gz     BQ4-018-M_S51_MERGED.fastq.gz 
#>                       0.005525345                       0.005606490 
#>    BQ9ABM-002_S52_MERGED.fastq.gz       BR8-005_S53_MERGED.fastq.gz 
#>                       0.005532056                       0.005525636 
#>      BS14-006_S54_MERGED.fastq.gz      BT-006-M_S55_MERGED.fastq.gz 
#>                       0.005417430                       0.005198680 
#>       BT7-006_S56_MERGED.fastq.gz    BV11-002-B_S57_MERGED.fastq.gz 
#>                       0.005113717                       0.005492216 
#>    BV11-002-H_S58_MERGED.fastq.gz    BV11-002-M_S59_MERGED.fastq.gz 
#>                       0.005472138                       0.005632221 
#>       BW8-003_S60_MERGED.fastq.gz        C1-001_S61_MERGED.fastq.gz 
#>                       0.005379245                       0.005308420 
#>     C21-NV1-B_S62_MERGED.fastq.gz     C21-NV1-H_S63_MERGED.fastq.gz 
#>                       0.005545973                       0.005511839 
#>     C21-NV1-M_S64_MERGED.fastq.gz        C9-005_S65_MERGED.fastq.gz 
#>                       0.005743994                       0.005625365 
#>      CA12-024_S66_MERGED.fastq.gz       CA9-027_S67_MERGED.fastq.gz 
#>                       0.005587228                       0.005743050 
#>         CA9-X_S68_MERGED.fastq.gz     CB8-019-B_S69_MERGED.fastq.gz 
#>                       0.005460232                       0.005606963 
#>     CB8-019-H_S70_MERGED.fastq.gz     CB8-019-M_S71_MERGED.fastq.gz 
#>                       0.005536705                       0.005564729 
#>       CB9-013_S72_MERGED.fastq.gz       CC3-044_S73_MERGED.fastq.gz 
#>                       0.005524410                       0.005491664 
#>       CC8-003_S74_MERGED.fastq.gz       D17-011_S77_MERGED.fastq.gz 
#>                       0.005636944                       0.005390811 
#>     D18-003-B_S78_MERGED.fastq.gz     D18-003-H_S79_MERGED.fastq.gz 
#>                       0.005491481                       0.005557191 
#>     D18-003-M_S80_MERGED.fastq.gz    D22-NVABM1_S81_MERGED.fastq.gz 
#>                       0.005613003                       0.005399096 
#>     D61-010-B_S82_MERGED.fastq.gz      D9-027-B_S83_MERGED.fastq.gz 
#>                       0.005525214                       0.005243712 
#>      D9-027-H_S84_MERGED.fastq.gz      D9-027-M_S85_MERGED.fastq.gz 
#>                       0.005057125                       0.005270961 
#>   DBM-ABM-001_S86_MERGED.fastq.gz     DJ2-008-B_S87_MERGED.fastq.gz 
#>                       0.005344133                       0.005432589 
#>     DJ2-008-H_S88_MERGED.fastq.gz     DJ2-008-M_S89_MERGED.fastq.gz 
#>                       0.005667010                       0.005312133 
#>    DP4-ABM001_S90_MERGED.fastq.gz  DS1-ABM002-B_S91_MERGED.fastq.gz 
#>                       0.005600380                       0.005512753 
#>  DS1-ABM002-H_S92_MERGED.fastq.gz  DS1-ABM002-M_S93_MERGED.fastq.gz 
#>                       0.005689445                       0.005526927 
#>     DU3-045-B_S94_MERGED.fastq.gz       DW4-007_S95_MERGED.fastq.gz 
#>                       0.005380856                       0.005500723 
#>     DY5-004-B_S96_MERGED.fastq.gz     DY5-004-H_S97_MERGED.fastq.gz 
#>                       0.005182943                       0.005088777 
#>     DY5-004-M_S98_MERGED.fastq.gz   DZ6-ABM-001_S99_MERGED.fastq.gz 
#>                       0.005203781                       0.005593447 
#>     E9-009-B_S100_MERGED.fastq.gz     E9-009-H_S101_MERGED.fastq.gz 
#>                       0.005093335                       0.005052157 
#>     E9-009-M_S102_MERGED.fastq.gz  EA5-ABM-001_S103_MERGED.fastq.gz 
#>                       0.004981648                       0.005515744 
#>    EC2-013-B_S104_MERGED.fastq.gz   F6-ABM-001_S105_MERGED.fastq.gz 
#>                       0.005582450                       0.005767824 
#>     F7-015-M_S106_MERGED.fastq.gz    FOMES19-H_S108_MERGED.fastq.gz 
#>                       0.005757017                       0.005119403 
#>    FOMES19-M_S109_MERGED.fastq.gz    H10-018-M_S110_MERGED.fastq.gz 
#>                       0.005237816                       0.005572858 
#> H24-NVABM1-H_S111_MERGED.fastq.gz    J18-004-B_S114_MERGED.fastq.gz 
#>                       0.005271013                       0.005182117 
#>    J18-004-H_S115_MERGED.fastq.gz    J18-004-M_S116_MERGED.fastq.gz 
#>                       0.005799836                       0.005164878 
#>    K18-002-H_S117_MERGED.fastq.gz   K26-NVABM1_S118_MERGED.fastq.gz 
#>                       0.005525452                       0.005536350 
#>       L19X-B_S119_MERGED.fastq.gz       L19X-H_S120_MERGED.fastq.gz 
#>                       0.005776322                       0.005550468 
#>       L19X-M_S121_MERGED.fastq.gz    L23-002-B_S122_MERGED.fastq.gz 
#>                       0.005397867                       0.005173481 
#>    L23-002-H_S123_MERGED.fastq.gz    L23-002-M_S124_MERGED.fastq.gz 
#>                       0.005607710                       0.005288520 
#>      M22-001_S125_MERGED.fastq.gz       N19X-B_S126_MERGED.fastq.gz 
#>                       0.005620118                       0.005069924 
#>       N19X-H_S127_MERGED.fastq.gz       N19X-M_S128_MERGED.fastq.gz 
#>                       0.005296358                       0.005410732 
#>    N22-001-B_S129_MERGED.fastq.gz    N23-002-B_S130_MERGED.fastq.gz 
#>                       0.005797441                       0.005504839 
#>    N23-002-H_S131_MERGED.fastq.gz     N25-ABMX_S133_MERGED.fastq.gz 
#>                       0.005505184                       0.005570539 
#>   NVABM-0058_S134_MERGED.fastq.gz NVABM-0163-H_S135_MERGED.fastq.gz 
#>                       0.005498004                       0.005552250 
#>   NVABM-0397_S138_MERGED.fastq.gz    NVABM0216_S136_MERGED.fastq.gz 
#>                       0.005541689                       0.005670066 
#>  NVABM0244-M_S137_MERGED.fastq.gz      O20-X-B_S139_MERGED.fastq.gz 
#>                       0.005753293                       0.005585280 
#>      O20-X-H_S140_MERGED.fastq.gz      O20-X-M_S141_MERGED.fastq.gz 
#>                       0.005246993                       0.005237162 
#>    O21-007-B_S142_MERGED.fastq.gz    O21-007-H_S143_MERGED.fastq.gz 
#>                       0.005499893                       0.005608886 
#>    O21-007-M_S144_MERGED.fastq.gz    O24-003-B_S145_MERGED.fastq.gz 
#>                       0.005972513                       0.005346457 
#>    O24-003-H_S146_MERGED.fastq.gz    O24-003-M_S147_MERGED.fastq.gz 
#>                       0.005502025                       0.005328186 
#>    O26-004-B_S148_MERGED.fastq.gz    O26-004-H_S149_MERGED.fastq.gz 
#>                       0.005598397                       0.005573576 
#>    O26-004-M_S150_MERGED.fastq.gz      O27-012_S151_MERGED.fastq.gz 
#>                       0.005774263                       0.005621182 
#>     O9-005-B_S152_MERGED.fastq.gz    P19-023-M_S153_MERGED.fastq.gz 
#>                       0.005121205                       0.005585693 
#>    P27-015-M_S154_MERGED.fastq.gz   P27-ABM001_S155_MERGED.fastq.gz 
#>                       0.005654745                       0.005672433 
#> Q27-ABM003-B_S156_MERGED.fastq.gz     R25-ABMX_S157_MERGED.fastq.gz 
#>                       0.005491006                       0.005448344 
#>    R28-008-B_S158_MERGED.fastq.gz    R28-008-H_S159_MERGED.fastq.gz 
#>                       0.005131425                       0.005015862 
#>    R28-008-M_S160_MERGED.fastq.gz      T28-011_S161_MERGED.fastq.gz 
#>                       0.005106585                       0.005512879 
#> T28-ABM602-B_S162_MERGED.fastq.gz   U27-ABM002_S163_MERGED.fastq.gz 
#>                       0.005813927                       0.005403655 
#>     W25-ABMX_S164_MERGED.fastq.gz    W26-001-B_S165_MERGED.fastq.gz 
#>                       0.005486949                       0.004783710 
#>    W26-001-H_S166_MERGED.fastq.gz    W26-001-M_S167_MERGED.fastq.gz 
#>                       0.004935947                       0.005094514 
#>      W30-006_S168_MERGED.fastq.gz     W9-025-M_S169_MERGED.fastq.gz 
#>                       0.005427852                       0.005765172 
#>    X24-009-B_S170_MERGED.fastq.gz    X24-009-H_S171_MERGED.fastq.gz 
#>                       0.005299386                       0.005621477 
#>    X24-009-M_S172_MERGED.fastq.gz      X24-010_S173_MERGED.fastq.gz 
#>                       0.005583280                       0.005599091 
#>    X29-004-B_S174_MERGED.fastq.gz    X29-004-H_S175_MERGED.fastq.gz 
#>                       0.005693585                       0.005305343 
#>    X29-004-M_S176_MERGED.fastq.gz Y21-ABM484-H_S177_MERGED.fastq.gz 
#>                       0.005500181                       0.005406702 
#>    Y28-002-B_S178_MERGED.fastq.gz    Y28-002-H_S179_MERGED.fastq.gz 
#>                       0.005310932                       0.005155402 
#>    Y28-002-M_S180_MERGED.fastq.gz    Y29-007-B_S181_MERGED.fastq.gz 
#>                       0.005371721                       0.005157384 
#>    Y29-007-H_S182_MERGED.fastq.gz    Y29-007-M_S183_MERGED.fastq.gz 
#>                       0.005124434                       0.005539944 
#> Y31-ABM484-B_S184_MERGED.fastq.gz    Z29-001-H_S185_MERGED.fastq.gz 
#>                       0.005617242                       0.005465091 
#>      Z30-002_S186_MERGED.fastq.gz Z30-ABM560-M_S187_MERGED.fastq.gz 
#>                       0.005118409                       0.005401336 
#> 
#> $p.LCBD
#>    A10-005-B_S188_MERGED.fastq.gz    A10-005-H_S189_MERGED.fastq.gz 
#>                         0.3333333                         0.3333333 
#>    A10-005-M_S190_MERGED.fastq.gz      A12-007_S191_MERGED.fastq.gz 
#>                         0.5000000                         0.1666667 
#>      A12-007-B_S2_MERGED.fastq.gz        A15-004_S3_MERGED.fastq.gz 
#>                         0.1666667                         1.0000000 
#>         A8-005_S4_MERGED.fastq.gz         A9-012_S5_MERGED.fastq.gz 
#>                         0.5000000                         0.8333333 
#>    AB29-ABMX-H_S6_MERGED.fastq.gz       AC27-013_S7_MERGED.fastq.gz 
#>                         0.1666667                         0.6666667 
#>        AC29033_S8_MERGED.fastq.gz     AD26-005-B_S9_MERGED.fastq.gz 
#>                         0.8333333                         0.3333333 
#>    AD26-005-H_S10_MERGED.fastq.gz    AD26-005-M_S11_MERGED.fastq.gz 
#>                         0.3333333                         0.1666667 
#>   AD30-ABMX-M_S12_MERGED.fastq.gz    AD32-007-M_S13_MERGED.fastq.gz 
#>                         0.3333333                         0.3333333 
#>    ADABM30X-B_S14_MERGED.fastq.gz    ADABM30X-H_S15_MERGED.fastq.gz 
#>                         0.5000000                         0.6666667 
#>    ADABM30X-M_S16_MERGED.fastq.gz   AE30-ABM507_S17_MERGED.fastq.gz 
#>                         0.1666667                         0.5000000 
#>       B17-014_S18_MERGED.fastq.gz     B18-006-B_S19_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>   BA16-036bis_S20_MERGED.fastq.gz    BA17-050-B_S21_MERGED.fastq.gz 
#>                         0.1666667                         0.8333333 
#>    BB19-006-H_S22_MERGED.fastq.gz     BB6-019-B_S23_MERGED.fastq.gz 
#>                         0.5000000                         0.5000000 
#>     BB6-019-H_S24_MERGED.fastq.gz     BB6-019-M_S25_MERGED.fastq.gz 
#>                         0.1666667                         0.5000000 
#>      BD14-021_S26_MERGED.fastq.gz     BE9-006-B_S27_MERGED.fastq.gz 
#>                         1.0000000                         0.5000000 
#>     BE9-006-H_S28_MERGED.fastq.gz     BE9-006-M_S29_MERGED.fastq.gz 
#>                         0.8333333                         0.5000000 
#>     BG7-010-B_S30_MERGED.fastq.gz     BG7-010-H_S31_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     BG7-010-M_S32_MERGED.fastq.gz       BH9-021_S33_MERGED.fastq.gz 
#>                         1.0000000                         0.8333333 
#>    BJ17-007-M_S34_MERGED.fastq.gz   BJ8-ABM-003_S35_MERGED.fastq.gz 
#>                         0.6666667                         0.1666667 
#>     BL7-006-B_S36_MERGED.fastq.gz     BL7-006-H_S37_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     BL7-006-M_S38_MERGED.fastq.gz      BN11-041_S39_MERGED.fastq.gz 
#>                         1.0000000                         0.5000000 
#>       BN9-002_S40_MERGED.fastq.gz       BO8-002_S41_MERGED.fastq.gz 
#>                         0.3333333                         0.3333333 
#>       BO8-005_S42_MERGED.fastq.gz    BP11-001-B_S43_MERGED.fastq.gz 
#>                         0.5000000                         0.3333333 
#>    BP11-001-H_S44_MERGED.fastq.gz    BP11-001-M_S45_MERGED.fastq.gz 
#>                         0.8333333                         0.5000000 
#>    BP12-025-B_S46_MERGED.fastq.gz      BP14-006_S47_MERGED.fastq.gz 
#>                         0.5000000                         0.8333333 
#>       BQ3-019_S48_MERGED.fastq.gz     BQ4-018-B_S49_MERGED.fastq.gz 
#>                         0.1666667                         0.6666667 
#>     BQ4-018-H_S50_MERGED.fastq.gz     BQ4-018-M_S51_MERGED.fastq.gz 
#>                         0.5000000                         0.3333333 
#>    BQ9ABM-002_S52_MERGED.fastq.gz       BR8-005_S53_MERGED.fastq.gz 
#>                         0.3333333                         0.6666667 
#>      BS14-006_S54_MERGED.fastq.gz      BT-006-M_S55_MERGED.fastq.gz 
#>                         0.6666667                         1.0000000 
#>       BT7-006_S56_MERGED.fastq.gz    BV11-002-B_S57_MERGED.fastq.gz 
#>                         1.0000000                         0.5000000 
#>    BV11-002-H_S58_MERGED.fastq.gz    BV11-002-M_S59_MERGED.fastq.gz 
#>                         0.3333333                         0.1666667 
#>       BW8-003_S60_MERGED.fastq.gz        C1-001_S61_MERGED.fastq.gz 
#>                         0.8333333                         1.0000000 
#>     C21-NV1-B_S62_MERGED.fastq.gz     C21-NV1-H_S63_MERGED.fastq.gz 
#>                         0.5000000                         0.3333333 
#>     C21-NV1-M_S64_MERGED.fastq.gz        C9-005_S65_MERGED.fastq.gz 
#>                         0.1666667                         0.3333333 
#>      CA12-024_S66_MERGED.fastq.gz       CA9-027_S67_MERGED.fastq.gz 
#>                         0.1666667                         0.1666667 
#>         CA9-X_S68_MERGED.fastq.gz     CB8-019-B_S69_MERGED.fastq.gz 
#>                         0.5000000                         0.1666667 
#>     CB8-019-H_S70_MERGED.fastq.gz     CB8-019-M_S71_MERGED.fastq.gz 
#>                         0.5000000                         0.3333333 
#>       CB9-013_S72_MERGED.fastq.gz       CC3-044_S73_MERGED.fastq.gz 
#>                         0.3333333                         0.5000000 
#>       CC8-003_S74_MERGED.fastq.gz       D17-011_S77_MERGED.fastq.gz 
#>                         0.1666667                         0.6666667 
#>     D18-003-B_S78_MERGED.fastq.gz     D18-003-H_S79_MERGED.fastq.gz 
#>                         0.3333333                         0.3333333 
#>     D18-003-M_S80_MERGED.fastq.gz    D22-NVABM1_S81_MERGED.fastq.gz 
#>                         0.1666667                         0.3333333 
#>     D61-010-B_S82_MERGED.fastq.gz      D9-027-B_S83_MERGED.fastq.gz 
#>                         0.3333333                         1.0000000 
#>      D9-027-H_S84_MERGED.fastq.gz      D9-027-M_S85_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>   DBM-ABM-001_S86_MERGED.fastq.gz     DJ2-008-B_S87_MERGED.fastq.gz 
#>                         1.0000000                         0.5000000 
#>     DJ2-008-H_S88_MERGED.fastq.gz     DJ2-008-M_S89_MERGED.fastq.gz 
#>                         0.1666667                         0.5000000 
#>    DP4-ABM001_S90_MERGED.fastq.gz  DS1-ABM002-B_S91_MERGED.fastq.gz 
#>                         0.3333333                         0.1666667 
#>  DS1-ABM002-H_S92_MERGED.fastq.gz  DS1-ABM002-M_S93_MERGED.fastq.gz 
#>                         0.1666667                         0.3333333 
#>     DU3-045-B_S94_MERGED.fastq.gz       DW4-007_S95_MERGED.fastq.gz 
#>                         0.6666667                         0.3333333 
#>     DY5-004-B_S96_MERGED.fastq.gz     DY5-004-H_S97_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     DY5-004-M_S98_MERGED.fastq.gz   DZ6-ABM-001_S99_MERGED.fastq.gz 
#>                         1.0000000                         0.1666667 
#>     E9-009-B_S100_MERGED.fastq.gz     E9-009-H_S101_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     E9-009-M_S102_MERGED.fastq.gz  EA5-ABM-001_S103_MERGED.fastq.gz 
#>                         1.0000000                         0.5000000 
#>    EC2-013-B_S104_MERGED.fastq.gz   F6-ABM-001_S105_MERGED.fastq.gz 
#>                         0.1666667                         0.1666667 
#>     F7-015-M_S106_MERGED.fastq.gz    FOMES19-H_S108_MERGED.fastq.gz 
#>                         0.1666667                         1.0000000 
#>    FOMES19-M_S109_MERGED.fastq.gz    H10-018-M_S110_MERGED.fastq.gz 
#>                         1.0000000                         0.5000000 
#> H24-NVABM1-H_S111_MERGED.fastq.gz    J18-004-B_S114_MERGED.fastq.gz 
#>                         0.8333333                         1.0000000 
#>    J18-004-H_S115_MERGED.fastq.gz    J18-004-M_S116_MERGED.fastq.gz 
#>                         0.1666667                         1.0000000 
#>    K18-002-H_S117_MERGED.fastq.gz   K26-NVABM1_S118_MERGED.fastq.gz 
#>                         0.5000000                         0.3333333 
#>       L19X-B_S119_MERGED.fastq.gz       L19X-H_S120_MERGED.fastq.gz 
#>                         0.1666667                         0.5000000 
#>       L19X-M_S121_MERGED.fastq.gz    L23-002-B_S122_MERGED.fastq.gz 
#>                         0.6666667                         1.0000000 
#>    L23-002-H_S123_MERGED.fastq.gz    L23-002-M_S124_MERGED.fastq.gz 
#>                         0.1666667                         1.0000000 
#>      M22-001_S125_MERGED.fastq.gz       N19X-B_S126_MERGED.fastq.gz 
#>                         0.3333333                         1.0000000 
#>       N19X-H_S127_MERGED.fastq.gz       N19X-M_S128_MERGED.fastq.gz 
#>                         1.0000000                         0.6666667 
#>    N22-001-B_S129_MERGED.fastq.gz    N23-002-B_S130_MERGED.fastq.gz 
#>                         0.1666667                         0.3333333 
#>    N23-002-H_S131_MERGED.fastq.gz     N25-ABMX_S133_MERGED.fastq.gz 
#>                         0.8333333                         0.1666667 
#>   NVABM-0058_S134_MERGED.fastq.gz NVABM-0163-H_S135_MERGED.fastq.gz 
#>                         0.6666667                         0.1666667 
#>   NVABM-0397_S138_MERGED.fastq.gz    NVABM0216_S136_MERGED.fastq.gz 
#>                         0.5000000                         0.1666667 
#>  NVABM0244-M_S137_MERGED.fastq.gz      O20-X-B_S139_MERGED.fastq.gz 
#>                         0.1666667                         0.3333333 
#>      O20-X-H_S140_MERGED.fastq.gz      O20-X-M_S141_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    O21-007-B_S142_MERGED.fastq.gz    O21-007-H_S143_MERGED.fastq.gz 
#>                         0.1666667                         0.1666667 
#>    O21-007-M_S144_MERGED.fastq.gz    O24-003-B_S145_MERGED.fastq.gz 
#>                         0.1666667                         0.8333333 
#>    O24-003-H_S146_MERGED.fastq.gz    O24-003-M_S147_MERGED.fastq.gz 
#>                         0.3333333                         1.0000000 
#>    O26-004-B_S148_MERGED.fastq.gz    O26-004-H_S149_MERGED.fastq.gz 
#>                         0.1666667                         0.1666667 
#>    O26-004-M_S150_MERGED.fastq.gz      O27-012_S151_MERGED.fastq.gz 
#>                         0.1666667                         0.3333333 
#>     O9-005-B_S152_MERGED.fastq.gz    P19-023-M_S153_MERGED.fastq.gz 
#>                         1.0000000                         0.1666667 
#>    P27-015-M_S154_MERGED.fastq.gz   P27-ABM001_S155_MERGED.fastq.gz 
#>                         0.1666667                         0.1666667 
#> Q27-ABM003-B_S156_MERGED.fastq.gz     R25-ABMX_S157_MERGED.fastq.gz 
#>                         0.3333333                         0.5000000 
#>    R28-008-B_S158_MERGED.fastq.gz    R28-008-H_S159_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    R28-008-M_S160_MERGED.fastq.gz      T28-011_S161_MERGED.fastq.gz 
#>                         1.0000000                         0.3333333 
#> T28-ABM602-B_S162_MERGED.fastq.gz   U27-ABM002_S163_MERGED.fastq.gz 
#>                         0.1666667                         0.5000000 
#>     W25-ABMX_S164_MERGED.fastq.gz    W26-001-B_S165_MERGED.fastq.gz 
#>                         0.3333333                         1.0000000 
#>    W26-001-H_S166_MERGED.fastq.gz    W26-001-M_S167_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>      W30-006_S168_MERGED.fastq.gz     W9-025-M_S169_MERGED.fastq.gz 
#>                         0.3333333                         0.1666667 
#>    X24-009-B_S170_MERGED.fastq.gz    X24-009-H_S171_MERGED.fastq.gz 
#>                         0.8333333                         0.1666667 
#>    X24-009-M_S172_MERGED.fastq.gz      X24-010_S173_MERGED.fastq.gz 
#>                         0.1666667                         0.1666667 
#>    X29-004-B_S174_MERGED.fastq.gz    X29-004-H_S175_MERGED.fastq.gz 
#>                         0.1666667                         0.5000000 
#>    X29-004-M_S176_MERGED.fastq.gz Y21-ABM484-H_S177_MERGED.fastq.gz 
#>                         0.8333333                         0.6666667 
#>    Y28-002-B_S178_MERGED.fastq.gz    Y28-002-H_S179_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    Y28-002-M_S180_MERGED.fastq.gz    Y29-007-B_S181_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    Y29-007-H_S182_MERGED.fastq.gz    Y29-007-M_S183_MERGED.fastq.gz 
#>                         1.0000000                         0.1666667 
#> Y31-ABM484-B_S184_MERGED.fastq.gz    Z29-001-H_S185_MERGED.fastq.gz 
#>                         0.1666667                         0.6666667 
#>      Z30-002_S186_MERGED.fastq.gz Z30-ABM560-M_S187_MERGED.fastq.gz 
#>                         1.0000000                         0.8333333 
#> 
#> $p.adj
#>    A10-005-B_S188_MERGED.fastq.gz    A10-005-H_S189_MERGED.fastq.gz 
#>                         0.7572016                         0.7572016 
#>    A10-005-M_S190_MERGED.fastq.gz      A12-007_S191_MERGED.fastq.gz 
#>                         0.8363636                         0.6388889 
#>      A12-007-B_S2_MERGED.fastq.gz        A15-004_S3_MERGED.fastq.gz 
#>                         0.6388889                         1.0000000 
#>         A8-005_S4_MERGED.fastq.gz         A9-012_S5_MERGED.fastq.gz 
#>                         0.8363636                         1.0000000 
#>    AB29-ABMX-H_S6_MERGED.fastq.gz       AC27-013_S7_MERGED.fastq.gz 
#>                         0.6388889                         0.9972900 
#>        AC29033_S8_MERGED.fastq.gz     AD26-005-B_S9_MERGED.fastq.gz 
#>                         1.0000000                         0.7572016 
#>    AD26-005-H_S10_MERGED.fastq.gz    AD26-005-M_S11_MERGED.fastq.gz 
#>                         0.7572016                         0.6388889 
#>   AD30-ABMX-M_S12_MERGED.fastq.gz    AD32-007-M_S13_MERGED.fastq.gz 
#>                         0.7572016                         0.7572016 
#>    ADABM30X-B_S14_MERGED.fastq.gz    ADABM30X-H_S15_MERGED.fastq.gz 
#>                         0.8363636                         0.9972900 
#>    ADABM30X-M_S16_MERGED.fastq.gz   AE30-ABM507_S17_MERGED.fastq.gz 
#>                         0.6388889                         0.8363636 
#>       B17-014_S18_MERGED.fastq.gz     B18-006-B_S19_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>   BA16-036bis_S20_MERGED.fastq.gz    BA17-050-B_S21_MERGED.fastq.gz 
#>                         0.6388889                         1.0000000 
#>    BB19-006-H_S22_MERGED.fastq.gz     BB6-019-B_S23_MERGED.fastq.gz 
#>                         0.8363636                         0.8363636 
#>     BB6-019-H_S24_MERGED.fastq.gz     BB6-019-M_S25_MERGED.fastq.gz 
#>                         0.6388889                         0.8363636 
#>      BD14-021_S26_MERGED.fastq.gz     BE9-006-B_S27_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#>     BE9-006-H_S28_MERGED.fastq.gz     BE9-006-M_S29_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#>     BG7-010-B_S30_MERGED.fastq.gz     BG7-010-H_S31_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     BG7-010-M_S32_MERGED.fastq.gz       BH9-021_S33_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    BJ17-007-M_S34_MERGED.fastq.gz   BJ8-ABM-003_S35_MERGED.fastq.gz 
#>                         0.9972900                         0.6388889 
#>     BL7-006-B_S36_MERGED.fastq.gz     BL7-006-H_S37_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     BL7-006-M_S38_MERGED.fastq.gz      BN11-041_S39_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#>       BN9-002_S40_MERGED.fastq.gz       BO8-002_S41_MERGED.fastq.gz 
#>                         0.7572016                         0.7572016 
#>       BO8-005_S42_MERGED.fastq.gz    BP11-001-B_S43_MERGED.fastq.gz 
#>                         0.8363636                         0.7572016 
#>    BP11-001-H_S44_MERGED.fastq.gz    BP11-001-M_S45_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#>    BP12-025-B_S46_MERGED.fastq.gz      BP14-006_S47_MERGED.fastq.gz 
#>                         0.8363636                         1.0000000 
#>       BQ3-019_S48_MERGED.fastq.gz     BQ4-018-B_S49_MERGED.fastq.gz 
#>                         0.6388889                         0.9972900 
#>     BQ4-018-H_S50_MERGED.fastq.gz     BQ4-018-M_S51_MERGED.fastq.gz 
#>                         0.8363636                         0.7572016 
#>    BQ9ABM-002_S52_MERGED.fastq.gz       BR8-005_S53_MERGED.fastq.gz 
#>                         0.7572016                         0.9972900 
#>      BS14-006_S54_MERGED.fastq.gz      BT-006-M_S55_MERGED.fastq.gz 
#>                         0.9972900                         1.0000000 
#>       BT7-006_S56_MERGED.fastq.gz    BV11-002-B_S57_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#>    BV11-002-H_S58_MERGED.fastq.gz    BV11-002-M_S59_MERGED.fastq.gz 
#>                         0.7572016                         0.6388889 
#>       BW8-003_S60_MERGED.fastq.gz        C1-001_S61_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     C21-NV1-B_S62_MERGED.fastq.gz     C21-NV1-H_S63_MERGED.fastq.gz 
#>                         0.8363636                         0.7572016 
#>     C21-NV1-M_S64_MERGED.fastq.gz        C9-005_S65_MERGED.fastq.gz 
#>                         0.6388889                         0.7572016 
#>      CA12-024_S66_MERGED.fastq.gz       CA9-027_S67_MERGED.fastq.gz 
#>                         0.6388889                         0.6388889 
#>         CA9-X_S68_MERGED.fastq.gz     CB8-019-B_S69_MERGED.fastq.gz 
#>                         0.8363636                         0.6388889 
#>     CB8-019-H_S70_MERGED.fastq.gz     CB8-019-M_S71_MERGED.fastq.gz 
#>                         0.8363636                         0.7572016 
#>       CB9-013_S72_MERGED.fastq.gz       CC3-044_S73_MERGED.fastq.gz 
#>                         0.7572016                         0.8363636 
#>       CC8-003_S74_MERGED.fastq.gz       D17-011_S77_MERGED.fastq.gz 
#>                         0.6388889                         0.9972900 
#>     D18-003-B_S78_MERGED.fastq.gz     D18-003-H_S79_MERGED.fastq.gz 
#>                         0.7572016                         0.7572016 
#>     D18-003-M_S80_MERGED.fastq.gz    D22-NVABM1_S81_MERGED.fastq.gz 
#>                         0.6388889                         0.7572016 
#>     D61-010-B_S82_MERGED.fastq.gz      D9-027-B_S83_MERGED.fastq.gz 
#>                         0.7572016                         1.0000000 
#>      D9-027-H_S84_MERGED.fastq.gz      D9-027-M_S85_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>   DBM-ABM-001_S86_MERGED.fastq.gz     DJ2-008-B_S87_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#>     DJ2-008-H_S88_MERGED.fastq.gz     DJ2-008-M_S89_MERGED.fastq.gz 
#>                         0.6388889                         0.8363636 
#>    DP4-ABM001_S90_MERGED.fastq.gz  DS1-ABM002-B_S91_MERGED.fastq.gz 
#>                         0.7572016                         0.6388889 
#>  DS1-ABM002-H_S92_MERGED.fastq.gz  DS1-ABM002-M_S93_MERGED.fastq.gz 
#>                         0.6388889                         0.7572016 
#>     DU3-045-B_S94_MERGED.fastq.gz       DW4-007_S95_MERGED.fastq.gz 
#>                         0.9972900                         0.7572016 
#>     DY5-004-B_S96_MERGED.fastq.gz     DY5-004-H_S97_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     DY5-004-M_S98_MERGED.fastq.gz   DZ6-ABM-001_S99_MERGED.fastq.gz 
#>                         1.0000000                         0.6388889 
#>     E9-009-B_S100_MERGED.fastq.gz     E9-009-H_S101_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>     E9-009-M_S102_MERGED.fastq.gz  EA5-ABM-001_S103_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#>    EC2-013-B_S104_MERGED.fastq.gz   F6-ABM-001_S105_MERGED.fastq.gz 
#>                         0.6388889                         0.6388889 
#>     F7-015-M_S106_MERGED.fastq.gz    FOMES19-H_S108_MERGED.fastq.gz 
#>                         0.6388889                         1.0000000 
#>    FOMES19-M_S109_MERGED.fastq.gz    H10-018-M_S110_MERGED.fastq.gz 
#>                         1.0000000                         0.8363636 
#> H24-NVABM1-H_S111_MERGED.fastq.gz    J18-004-B_S114_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    J18-004-H_S115_MERGED.fastq.gz    J18-004-M_S116_MERGED.fastq.gz 
#>                         0.6388889                         1.0000000 
#>    K18-002-H_S117_MERGED.fastq.gz   K26-NVABM1_S118_MERGED.fastq.gz 
#>                         0.8363636                         0.7572016 
#>       L19X-B_S119_MERGED.fastq.gz       L19X-H_S120_MERGED.fastq.gz 
#>                         0.6388889                         0.8363636 
#>       L19X-M_S121_MERGED.fastq.gz    L23-002-B_S122_MERGED.fastq.gz 
#>                         0.9972900                         1.0000000 
#>    L23-002-H_S123_MERGED.fastq.gz    L23-002-M_S124_MERGED.fastq.gz 
#>                         0.6388889                         1.0000000 
#>      M22-001_S125_MERGED.fastq.gz       N19X-B_S126_MERGED.fastq.gz 
#>                         0.7572016                         1.0000000 
#>       N19X-H_S127_MERGED.fastq.gz       N19X-M_S128_MERGED.fastq.gz 
#>                         1.0000000                         0.9972900 
#>    N22-001-B_S129_MERGED.fastq.gz    N23-002-B_S130_MERGED.fastq.gz 
#>                         0.6388889                         0.7572016 
#>    N23-002-H_S131_MERGED.fastq.gz     N25-ABMX_S133_MERGED.fastq.gz 
#>                         1.0000000                         0.6388889 
#>   NVABM-0058_S134_MERGED.fastq.gz NVABM-0163-H_S135_MERGED.fastq.gz 
#>                         0.9972900                         0.6388889 
#>   NVABM-0397_S138_MERGED.fastq.gz    NVABM0216_S136_MERGED.fastq.gz 
#>                         0.8363636                         0.6388889 
#>  NVABM0244-M_S137_MERGED.fastq.gz      O20-X-B_S139_MERGED.fastq.gz 
#>                         0.6388889                         0.7572016 
#>      O20-X-H_S140_MERGED.fastq.gz      O20-X-M_S141_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    O21-007-B_S142_MERGED.fastq.gz    O21-007-H_S143_MERGED.fastq.gz 
#>                         0.6388889                         0.6388889 
#>    O21-007-M_S144_MERGED.fastq.gz    O24-003-B_S145_MERGED.fastq.gz 
#>                         0.6388889                         1.0000000 
#>    O24-003-H_S146_MERGED.fastq.gz    O24-003-M_S147_MERGED.fastq.gz 
#>                         0.7572016                         1.0000000 
#>    O26-004-B_S148_MERGED.fastq.gz    O26-004-H_S149_MERGED.fastq.gz 
#>                         0.6388889                         0.6388889 
#>    O26-004-M_S150_MERGED.fastq.gz      O27-012_S151_MERGED.fastq.gz 
#>                         0.6388889                         0.7572016 
#>     O9-005-B_S152_MERGED.fastq.gz    P19-023-M_S153_MERGED.fastq.gz 
#>                         1.0000000                         0.6388889 
#>    P27-015-M_S154_MERGED.fastq.gz   P27-ABM001_S155_MERGED.fastq.gz 
#>                         0.6388889                         0.6388889 
#> Q27-ABM003-B_S156_MERGED.fastq.gz     R25-ABMX_S157_MERGED.fastq.gz 
#>                         0.7572016                         0.8363636 
#>    R28-008-B_S158_MERGED.fastq.gz    R28-008-H_S159_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    R28-008-M_S160_MERGED.fastq.gz      T28-011_S161_MERGED.fastq.gz 
#>                         1.0000000                         0.7572016 
#> T28-ABM602-B_S162_MERGED.fastq.gz   U27-ABM002_S163_MERGED.fastq.gz 
#>                         0.6388889                         0.8363636 
#>     W25-ABMX_S164_MERGED.fastq.gz    W26-001-B_S165_MERGED.fastq.gz 
#>                         0.7572016                         1.0000000 
#>    W26-001-H_S166_MERGED.fastq.gz    W26-001-M_S167_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>      W30-006_S168_MERGED.fastq.gz     W9-025-M_S169_MERGED.fastq.gz 
#>                         0.7572016                         0.6388889 
#>    X24-009-B_S170_MERGED.fastq.gz    X24-009-H_S171_MERGED.fastq.gz 
#>                         1.0000000                         0.6388889 
#>    X24-009-M_S172_MERGED.fastq.gz      X24-010_S173_MERGED.fastq.gz 
#>                         0.6388889                         0.6388889 
#>    X29-004-B_S174_MERGED.fastq.gz    X29-004-H_S175_MERGED.fastq.gz 
#>                         0.6388889                         0.8363636 
#>    X29-004-M_S176_MERGED.fastq.gz Y21-ABM484-H_S177_MERGED.fastq.gz 
#>                         1.0000000                         0.9972900 
#>    Y28-002-B_S178_MERGED.fastq.gz    Y28-002-H_S179_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    Y28-002-M_S180_MERGED.fastq.gz    Y29-007-B_S181_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#>    Y29-007-H_S182_MERGED.fastq.gz    Y29-007-M_S183_MERGED.fastq.gz 
#>                         1.0000000                         0.6388889 
#> Y31-ABM484-B_S184_MERGED.fastq.gz    Z29-001-H_S185_MERGED.fastq.gz 
#>                         0.6388889                         0.9972900 
#>      Z30-002_S186_MERGED.fastq.gz Z30-ABM560-M_S187_MERGED.fastq.gz 
#>                         1.0000000                         1.0000000 
#> 
#> $method
#> [1] "jaccard"      "sqrt.D=FALSE"
#> 
#> $note
#> [1] "Info -- D is Euclidean because beta.div outputs D[jk] = sqrt(1-S[jk])"
#> [2] "For this D functions, use beta.div with option sqrt.D=FALSE"          
#> 
#> $D
#> [1] NA
#> 
#> attr(,"class")
#> [1] "beta.div"
# }