Deprecated function(s) in the MiscMetabar package

These functions are provided for compatibility with older version of the MiscMetabar package. They may eventually be completely removed.

Usage

physeq_graph_test(...)

Arguments

...

Parameters to be passed to the modern version of the function

Details

graph_test_pq	now a synonym for physeq_graph_test

adonis_pq	now a synonym for adonis_phyloseq

clean_pq	now a synonym for clean_physeq

lulu_pq	now a synonym for lulu_phyloseq

circle_pq	now a synonym for otu_circle

biplot_pq	now a synonym for biplot_physeq

read_pq	now a synonym for read_phyloseq

write_pq	now a synonym for write_phyloseq

sankey_pq	now a synonym for sankey_phyloseq

summary_plot_pq	now a synonym for summary_plot_phyloseq

plot_edgeR_pq	now a synonym for plot_edgeR_phyloseq

plot_deseq2_pq	now a synonym for plot_deseq2_phyloseq

venn_pq	now a synonym for venn_phyloseq

ggvenn_pq	now a synonym for ggVenn_phyloseq

hill_tuckey_pq	now a synonym for hill_tuckey_phyloseq

hill_pq	now a synonym for hill_phyloseq

heat_tree_pq	now a synonym for physeq_heat_tree

compare_pairs_pq	now a synonym for multiple_share_bisamples