Permanova on a phyloseq object

A wrapper for the vegan::adonis2() function in the case of physeq object.

Usage

adonis_pq(physeq, formula, merge_sample_by = NULL)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

formula

(required) todo

merge_sample_by

a vector to determine which samples to merge using the speedyseq::merge_samples2() function. Need to be in physeq@sam_data

Value

The function returns an anova.cca result object with a new column for partial R^2. See ?adonis2 for more information.

Author

Adrien Taudière

Examples

data(enterotype)
clean_enterotype <- subset_samples(
  enterotype,
  !is.na(enterotype@sam_data$Enterotype)
)
adonis_pq(clean_enterotype, "SeqTech*Enterotype")
#> Permutation test for adonis under reduced model
#> Terms added sequentially (first to last)
#> Permutation: free
#> Number of permutations: 999
#> 
#> vegan::adonis2(formula = .formula, data = metadata)
#>                     Df SumOfSqs      R2        F Pr(>F)    
#> SeqTech              2   29.175 0.54055 242.2289  0.001 ***
#> Enterotype           2    8.651 0.16028  71.8250  0.001 ***
#> SeqTech:Enterotype   4    0.368 0.00683   1.5293  0.112    
#> Residual           262   15.778 0.29234                    
#> Total              270   53.972 1.00000                    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1