Clean phyloseq object by removing empty samples and taxa

In addition, this function check for discrepancy (and rename) between (i) taxa names in refseq, taxonomy table and otu_table and between (ii) sample names in sam_data and otu_table.

Usage

clean_pq(
  physeq,
  remove_empty_samples = TRUE,
  remove_empty_taxa = TRUE,
  clean_samples_names = TRUE,
  silent = FALSE,
  verbose = FALSE,
  force_taxa_as_columns = FALSE,
  force_taxa_as_rows = FALSE,
  reorder_asv = FALSE,
  rename_asv = FALSE,
  simplify_taxo = FALSE
)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

remove_empty_samples

(logical) Do you want to remove samples without sequences (this is done after removing empty taxa)

remove_empty_taxa

(logical) Do you want to remove taxa without sequences (this is done before removing empty samples)

clean_samples_names

(logical) Do you want to clean samples names?

silent

(logical) If true, no message are printing.

verbose

(logical) Additional informations in the message the verbose parameter overwrite the silent parameter.

force_taxa_as_columns

(logical) If true, if the taxa are rows transpose the otu_table and set taxa_are_rows to false

force_taxa_as_rows

(logical) If true, if the taxa are columns transpose the otu_table and set taxa_are_rows to true

reorder_asv

(logical) if TRUE the otu_table is ordered by the number of sequences of ASV (descending order). Default to FALSE.

rename_asv

(logical) if TRUE, ASV are renamed by their position in the OTU_table (asv_1, asv_2, ...). Default to FALSE. If rename ASV is true, the ASV names in verbose information can be misleading.

simplify_taxo

(logical) if TRUE, correct the taxonomy_table using the MiscMetabar::simplify_taxo() function

Value

A new phyloseq-class object