Compute paired distances among matrix (e.g. otu_table)

May be used to verify ecological distance among samples.

Usage

dist_bycol(x, y, method = "bray", nperm = 99, ...)

Arguments

x

(required) A first matrix.

y

(required) A second matrix.

method

(default: 'bray') the method to use internally in the vegdist function.

nperm

(int) The number of permutations to perform.

...

Others argument for vegan::vegdist function

Value

A list of length two : (i) a vector of observed distance ($obs) and (ii) a matrix of the distance after randomization ($null)

Note

the first column of the first matrix is compare to the first column of the second matrix, the second column of the first matrix is compare to the second column of the second matrix and so on.

See also

Author

Adrien Taudière