Convert phyloseq data to DESeq2 dds object

No testing is performed by this function. The phyloseq data is converted to the relevant DESeqDataSet object, which can then be tested in the negative binomial generalized linear model framework of the DESeq function in DESeq2 package. See the phyloseq-extensions tutorials for more details.

phyloseq_to_deseq2(physeq, design, ...)

Arguments

physeq
(required). phyloseq-class. Must have a sample_data component.

design
(required). A formula which specifies the design of the experiment, taking the form formula(~ x + y + z). That is, a formula with right-hand side only. By default, the functions in this package and DESeq2 will use the last variable in the formula (e.g. z) for presenting results (fold changes, etc.) and plotting. When considering your specification of experimental design, you will want to re-order the levels so that the NULL set is first. For example, the following line of code would ensure that Enterotype 1 is used as the reference sample class in tests by setting it to the first of the factor levels using the relevel function:

sample_data(entill)$Enterotype <- relevel(sample_data(entill)$Enterotype, "1")

...
(Optional). Additional named arguments passed to DESeqDataSetFromMatrix. Most users will not need to pass any additional arguments here. Most testing-related options should be provided in a following call to DESeq.

Value

A DESeqDataSet object.

See also