Convert phyloseq data to DESeq2 dds object

No testing is performed by this function. The phyloseq data is converted to the relevant DESeqDataSet object, which can then be tested in the negative binomial generalized linear model framework of the DESeq function in DESeq2 package. See the phyloseq-extensions tutorials for more details.

phyloseq_to_deseq2(physeq, design, ...)

Arguments

physeq
(required). phyloseq-class. Must have a sample_data component.

design
(required). A formula which specifies the design of the experiment, taking the form formula(~ x + y + z). That is, a formula with right-hand side only. By default, the functions in this package and DESeq2 will use the last variable in the formula (e.g. z) for presenting results (fold changes, etc.) and plotting. When considering your specification of experimental design, you will want to re-order the levels so that the NULL set is first. For example, the following line of code would ensure that Enterotype 1 is used as the reference sample class in tests by setting it to the first of the factor levels using the relevel function:

sample_data(entill)$Enterotype <- relevel(sample_data(entill)$Enterotype, "1")

...
(Optional). Additional named arguments parsed to DESeqDataSetFromMatrix. Most users will not need to pass any additional arguments here. Most testing-related options should be provided in a following call to DESeq.

Value

A DESeqDataSet object.

See also

Examples

 # Check out the vignette phyloseq-mixture-models for more details.
 # vignette("phyloseq-mixture-models")
 data(soilrep)
 phyloseq_to_deseq2(soilrep, ~warmed)
#> Le chargement a nécessité le package : DESeq2
#> converting counts to integer mode
#> class: DESeqDataSet 
#> dim: 16825 56 
#> metadata(1): version
#> assays(1): counts
#> rownames(16825): OTU_R0 OTU_R1 ... OTU_R9997 OTU_R9998
#> rowData names(0):
#> colnames(56): a_C026 a_C066 ... a_C161 a_C162
#> colData names(4): Treatment warmed clipped Sample