Plot DESeq2 results for a phyloseq or a DESeq2 object.

Plot DESeq2 results for a phyloseq or a DESeq2 object.

plot_deseq2_phyloseq(
  data,
  contrast = NULL,
  tax_table = NULL,
  alpha = 0.01,
  taxa = "Genus",
  select_taxa = "No",
  color_tax = "Phylum",
  taxDepth = NULL,
  verbose = TRUE,
  ...
)

Arguments

data
(required): a phyloseq-class or a DESeqDataSet-class object.

contrast
(required):This argument specifies what comparison to extract from the object to build a results table. See results man page for more details.

tax_table
: Required if data is a DESeqDataSet-class object. The taxonomic table used to find the taxa and color_taxa arguments. If data is a phyloseq-class object, data@tax_table is used.

alpha
(default = 0.01): the significance cutoff used for optimizing the independent filtering. If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.

taxa
(default = 'Genus'): taxonomic level of interest

select_taxa
(default = 'No'): logical vector to select taxa to plot

color_tax
(default = 'Phylum'): taxonomic level used for color or a color vector.

taxDepth
(default = NULL): Taxonomic depth to test for differential distribution among contrast. If Null the analysis is done at the OTU (i.e. Species) level. If not Null data need to be a phyloseq-class object.

verbose
: whether the function print some information during the computation

...
Additional arguments passed on to DESeq or ggplot

Value

A ggplot2 plot representing DESeq2 results

See also