Sankey plot of phyloseq-class object

Sankey plot of phyloseq-class object

sankey_phyloseq(
  physeq = NULL,
  fact = NULL,
  taxa = c(1:4),
  nbSeq = FALSE,
  min.prop.tax = 0,
  tax2remove = NULL,
  units = NULL,
  Symbol2sub = c("\\.", "-"),
  ...
)

Arguments

physeq
(required): a phyloseq-class object.

fact
(Optional): Name of the factor to cluster samples by modalities. Need to be in physeq@sam_data.

taxa
(Default: c(1:4)): a vector of taxonomic rank to plot

nbSeq
(Default: FALSE): Represent the number of sequences or the number of OTUs (nbSeq = FALSE). Note that ploting the number of sequences is slower.

min.prop.tax
(Default: 0): The minimum proportion for taxon to be ploted. EXPERIMENTAL. For the moment each links below the min.prop. tax is discard from the sankey network resulting in sometimes weird plot.

tax2remove
: a vector of taxonomic groups to remove from the analysis (e.g. c('Incertae sedis', 'unidentified'))

units
: character string describing physical units (if any) for Value

Symbol2sub
(default = c('\.', '-')): vector of symbol to delete in the taxonomy

...
Additional arguments passed on to sankeyNetwork

data("GlobalPatterns") #GlobalPatterns_Archaea <- subset_taxa(GlobalPatterns, GlobalPatterns@tax_table[, 1] == 'Archaea') #sankey_phyloseq(GlobalPatterns_Archaea) #sankey_phyloseq(GlobalPatterns_Archaea, fact = 'SampleType') sankey_phyloseq(GlobalPatterns, taxa = c(1:5), min.prop.tax = 0.01) sankey_phyloseq(GlobalPatterns, taxa = c(2:6), min.prop.tax = 0.01, nbSeq = TRUE)

Value

A sankeyNetwork plot representing the taxonomic distribution of OTUs or sequences. If fact is set, represent the distribution of the last taxonomic level in the modalities of fact

See also