Subset taxa using a taxa control (e.g. truffle root tips) through 3 methods.

Usage

subset_taxa_tax_control(
  physeq,
  taxa_distri,
  method = "mean",
  min_diff_for_cutoff = NULL
)

Arguments

physeq

(required): a phyloseq-class object obtained using the phyloseq package.

taxa_distri

(required) a vector of length equal to the number of samples with the number of sequences per samples for the taxa control

method

(default: "mean") a method to calculate the cut-off value. There is 6 available methods:

cutoff_seq: discard taxa with less than the number of sequence than taxa control,

cutoff_mixt: using mixture models,

cutoff_diff: using a minimum difference threshold (need the argument min_diff_for_cutoff)

min: the minimum of the three firsts methods

max: the maximum of the three firsts methods

mean: the mean of the three firsts methods

min_diff_for_cutoff

(int) argument for method cutoff_diff. Required if method is cutoff_diff, min, max or mean

Value

A new phyloseq-class object.

Author

Adrien Taudière

Examples

data(data_fungi)

subset_taxa_tax_control(data_fungi,
  as.numeric(data_fungi@otu_table[, 300]),
  min_diff_for_cutoff = 2
)
#> number of iterations= 5 
#> number of iterations= 3 
#> number of iterations= 4 
#> number of iterations= 19 
#> number of iterations= 7 
#> number of iterations= 12 
#> Error in stats::uniroot(f = f, lower = 1, upper = 1000) : 
#>   f.upper = f(upper) is NA
#> Warning: NAs introduced by coercion
#> number of iterations= 8 
#> number of iterations= 15 
#> number of iterations= 3 
#> number of iterations= 10 
#> number of iterations= 14 
#> number of iterations= 14 
#> number of iterations= 14 
#> number of iterations= 9 
#> number of iterations= 25 
#> number of iterations= 38 
#> number of iterations= 15 
#> number of iterations= 16 
#> number of iterations= 5 
#> number of iterations= 42 
#> number of iterations= 9 
#> number of iterations= 22 
#> number of iterations= 4 
#> number of iterations= 15 
#> number of iterations= 18 
#> number of iterations= 13 
#> number of iterations= 8 
#> number of iterations= 12 
#> number of iterations= 10 
#> number of iterations= 13 
#> number of iterations= 10 
#> number of iterations= 4 
#> number of iterations= 17 
#> number of iterations= 7 
#> number of iterations= 14 
#> number of iterations= 18 
#> One of the variances is going to zero;  trying new starting values.
#> number of iterations= 11 
#> number of iterations= 7 
#> number of iterations= 16 
#> number of iterations= 9 
#> number of iterations= 7 
#> number of iterations= 21 
#> number of iterations= 6 
#> number of iterations= 3 
#> number of iterations= 7 
#> number of iterations= 8 
#> number of iterations= 13 
#> number of iterations= 3 
#> number of iterations= 8 
#> number of iterations= 15 
#> number of iterations= 12 
#> number of iterations= 22 
#> number of iterations= 9 
#> number of iterations= 22 
#> number of iterations= 10 
#> number of iterations= 8 
#> number of iterations= 9 
#> number of iterations= 11 
#> number of iterations= 14 
#> number of iterations= 7 
#> number of iterations= 24 
#> number of iterations= 19 
#> number of iterations= 4 
#> number of iterations= 5 
#> number of iterations= 6 
#> number of iterations= 13 
#> number of iterations= 7 
#> Error in stats::uniroot(f = f, lower = 1, upper = 1000) : 
#>   f.upper = f(upper) is NA
#> Warning: NAs introduced by coercion
#> number of iterations= 8 
#> number of iterations= 5 
#> number of iterations= 19 
#> number of iterations= 13 
#> number of iterations= 15 
#> number of iterations= 7 
#> number of iterations= 3 
#> number of iterations= 5 
#> number of iterations= 18 
#> number of iterations= 10 
#> number of iterations= 8 
#> Error in stats::uniroot(f = f, lower = 1, upper = 1000) : 
#>   f.upper = f(upper) is NA
#> Warning: NAs introduced by coercion
#> number of iterations= 7 
#> number of iterations= 12 
#> number of iterations= 11 
#> number of iterations= 8 
#> number of iterations= 13 
#> number of iterations= 13 
#> number of iterations= 6 
#> number of iterations= 10 
#> number of iterations= 3 
#> number of iterations= 25 
#> Warning: no non-missing arguments to min; returning Inf
#> number of iterations= 9 
#> number of iterations= 19 
#> number of iterations= 8 
#> number of iterations= 18 
#> number of iterations= 12 
#> number of iterations= 16 
#> number of iterations= 4 
#> number of iterations= 8 
#> number of iterations= 14 
#> number of iterations= 23 
#> One of the variances is going to zero;  trying new starting values.
#> number of iterations= 10 
#> number of iterations= 5 
#> number of iterations= 4 
#> number of iterations= 25 
#> number of iterations= 7 
#> number of iterations= 6 
#> Error in stats::uniroot(f = f, lower = 1, upper = 1000) : 
#>   f.upper = f(upper) is NA
#> Warning: NAs introduced by coercion
#> number of iterations= 19 
#> number of iterations= 5 
#> number of iterations= 5 
#> number of iterations= 8 
#> number of iterations= 6 
#> number of iterations= 12 
#> number of iterations= 15 
#> number of iterations= 20 
#> number of iterations= 4 
#> number of iterations= 25 
#> number of iterations= 5 
#> number of iterations= 21 
#> number of iterations= 19 
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to min; returning Inf
#> number of iterations= 8 
#> Warning: no non-missing arguments to min; returning Inf
#> number of iterations= 8 
#> number of iterations= 17 
#> One of the variances is going to zero;  trying new starting values.
#> number of iterations= 5 
#> number of iterations= 8 
#> number of iterations= 15 
#> number of iterations= 5 
#> number of iterations= 5 
#> number of iterations= 10 
#> Warning: no non-missing arguments to min; returning Inf
#> number of iterations= 7 
#> number of iterations= 18 
#> number of iterations= 35 
#> number of iterations= 14 
#> number of iterations= 7 
#> number of iterations= 33 
#> number of iterations= 10 
#> number of iterations= 3 
#> number of iterations= 5 
#> number of iterations= 9 
#> number of iterations= 16 
#> number of iterations= 7 
#> number of iterations= 4 
#> number of iterations= 11 
#> number of iterations= 3 
#> number of iterations= 12 
#> number of iterations= 9 
#> The filtering processes discard 64 taxa and 36109 sequences
#> phyloseq-class experiment-level object
#> otu_table()   OTU Table:          [ 1356 taxa and 185 samples ]:
#> sample_data() Sample Data:        [ 185 samples by 7 sample variables ]:
#> tax_table()   Taxonomy Table:     [ 1356 taxa by 14 taxonomic ranks ]:
#> refseq()      DNAStringSet:       [ 1356 reference sequences ]
#> taxa are columns