Make a datatable with the taxonomy of a phyloseq-class object

Usage

tax_datatable(
  physeq,
  abundance = TRUE,
  taxonomic_level = NULL,
  modality = NULL,
  ...
)

Arguments

physeq

(required): A phyloseq-class object

abundance

(Default: TRUE): Does the number of sequences is print

taxonomic_level

(Default: NULL): a vector of selected taxonomic level using their column numbers (e.g. taxonomic_level = c(1:7))

modality

(Default: NULL) : A sample modality to split OTU abundancy by level of the modality

...

Other argument for the datatable function

Value

A datatable

Author

Adrien Taudière

Examples

data("GlobalPatterns")
tax_datatable(subset_taxa(
  GlobalPatterns,
  rowSums(GlobalPatterns@otu_table) > 10000
))

# Using modality
tax_datatable(GlobalPatterns,
  modality = GlobalPatterns@sam_data$SampleType
)