Track the number of reads (= sequences), samples and cluster (e.g. ASV) from various objects including dada-class and derep-class.

List of fastq and fastg.gz files -> nb of reads and samples

List of dada-class -> nb of reads, clusters (ASV) and samples

List of derep-class -> nb of reads, clusters (unique sequences) and samples

Matrix of samples x clusters (e.g. otu_table) -> nb of reads, clusters and samples

Phyloseq-class -> nb of reads, clusters and samples

Usage

track_wkflow(
  list_of_objects,
  obj_names = NULL,
  clean_pq = FALSE,
  taxonomy_rank = NULL,
  ...
)

Arguments

list_of_objects

(required) a list of objects

obj_names

A list of names corresponding to the list of objects

clean_pq

(logical) If set to TRUE, empty samples and empty ASV are discarded before clustering.

taxonomy_rank

A vector of int. Define the column number of taxonomic rank in physeq@tax_table to compute the number of unique value. Default is NULL and do not compute values for any taxonomic rank

...

Other arguments passed on to clean_pq() function.

Value

The number of sequences, clusters (e.g. OTUs, ASVs) and samples for each object.